Crate wgpu_core

This library safely implements WebGPU on native platforms. It is designed for integration into browsers, as well as wrapping into other language-specific user-friendly libraries.

Modules

api

binding_model

command

device

error

hub

Allocating resource ids, and tracking the resources they refer to.

id

instance

pipeline

present

Presentation.

resource

Macros

gfx_if_dx11_hidden

gfx_if_dx12_hidden

gfx_if_gles_hidden

gfx_if_metal_hidden

gfx_if_vulkan_hidden

gfx_select

Dispatch on an Id’s backend to a backend-generic method.

Structs

LifeGuard

Information needed to decide when it’s safe to free some wgpu-core resource.

Constants

MAX_BIND_GROUPS

MAX_COLOR_ATTACHMENTS

MAX_VERTEX_BUFFERS

Type Definitions

Label

RawString