wgpu_core::binding_model
Struct BindingTypeMaxCountValidator

pub(crate) struct BindingTypeMaxCountValidator {
    dynamic_uniform_buffers: u32,
    dynamic_storage_buffers: u32,
    sampled_textures: PerStageBindingTypeCounter,
    samplers: PerStageBindingTypeCounter,
    storage_buffers: PerStageBindingTypeCounter,
    storage_textures: PerStageBindingTypeCounter,
    uniform_buffers: PerStageBindingTypeCounter,
    acceleration_structures: PerStageBindingTypeCounter,
    binding_array_elements: PerStageBindingTypeCounter,
    binding_array_sampler_elements: PerStageBindingTypeCounter,
    has_bindless_array: bool,
}

Fields
dynamic_uniform_buffers: u32dynamic_storage_buffers: u32sampled_textures: PerStageBindingTypeCountersamplers: PerStageBindingTypeCounterstorage_buffers: PerStageBindingTypeCounterstorage_textures: PerStageBindingTypeCounteruniform_buffers: PerStageBindingTypeCounteracceleration_structures: PerStageBindingTypeCounterbinding_array_elements: PerStageBindingTypeCounterbinding_array_sampler_elements: PerStageBindingTypeCounterhas_bindless_array: bool
Implementations

impl BindingTypeMaxCountValidator

pub(crate) fn add_binding(&mut self, binding: &BindGroupLayoutEntry)

pub(crate) fn merge(&mut self, other: &Self)

pub(crate) fn validate( &self, limits: &Limits, ) -> Result<(), BindingTypeMaxCountError>

pub(crate) fn validate_binding_arrays( &self, ) -> Result<(), CreateBindGroupLayoutError>

Validate that the bind group layout does not contain both a binding array and a dynamic offset array.

This allows us to use UPDATE_AFTER_BIND on vulkan for bindless arrays. Vulkan does not allow UPDATE_AFTER_BIND on dynamic offset arrays. See https://github.com/gfx-rs/wgpu/issues/6737