Type Alias wgpu_core::binding_model::BindGroupLayouts

pub type BindGroupLayouts<A> = Storage<BindGroupLayout<A>, BindGroupLayoutId>;

Aliased Type

struct BindGroupLayouts<A> {
    pub(crate) map: Vec<Element<BindGroupLayout<A>>, Global>,
    pub(crate) kind: &'static str,
    pub(crate) _phantom: PhantomData<Id<BindGroupLayout<Api>>>,
}

Fields
map: Vec<Element<BindGroupLayout<A>>, Global>kind: &'static str_phantom: PhantomData<Id<BindGroupLayout<Api>>>
Implementations

impl<T, I: TypedId> Storage<T, I>

pub(crate) fn contains(&self, id: I) -> bool

pub(crate) fn try_get(&self, id: I) -> Result<Option<&T>, InvalidId>

Attempts to get a reference to an item behind a potentially invalid ID.

Returns None if there is an epoch mismatch, or the entry is empty.

This function is primarily intended for the as_hal family of functions where you may need to fallibly get a object backed by an id that could be in a different hub.

pub(crate) fn get(&self, id: I) -> Result<&T, InvalidId>

Get a reference to an item behind a potentially invalid ID. Panics if there is an epoch mismatch, or the entry is empty.

pub(crate) fn get_mut(&mut self, id: I) -> Result<&mut T, InvalidId>

Get a mutable reference to an item behind a potentially invalid ID. Panics if there is an epoch mismatch, or the entry is empty.

pub(crate) unsafe fn get_unchecked(&self, id: u32) -> &T

pub(crate) fn label_for_invalid_id(&self, id: I) -> &str

fn insert_impl(&mut self, index: usize, element: Element<T>)

pub(crate) fn insert(&mut self, id: I, value: T)

pub(crate) fn insert_error(&mut self, id: I, label: &str)

pub(crate) fn force_replace(&mut self, id: I, value: T)

pub(crate) fn remove(&mut self, id: I) -> Option<T>

pub(crate) fn _try_remove(&mut self, id: I) -> Option<T>

pub(crate) fn iter(&self, backend: Backend) -> impl Iterator<Item = (I, &T)>

pub(crate) fn len(&self) -> usize

pub(crate) fn generate_report(&self) -> StorageReport