wgpu_core::binding_model

Type Alias ResolvedPipelineLayoutDescriptor

pub type ResolvedPipelineLayoutDescriptor<'a, BGL = Arc<BindGroupLayout>> = PipelineLayoutDescriptor<'a, BGL>;

cbindgen:ignore

Aliased Type

struct ResolvedPipelineLayoutDescriptor<'a, BGL = Arc<BindGroupLayout>> {
    pub label: Option<Cow<'a, str>>,
    pub bind_group_layouts: Cow<'a, [BGL]>,
    pub immediate_size: u32,
}

Fields
label: Option<Cow<'a, str>>

Debug label of the pipeline layout.

This will show up in graphics debuggers for easy identification.

bind_group_layouts: Cow<'a, [BGL]>

Bind groups that this pipeline uses. The first entry will provide all the bindings for “set = 0”, second entry will provide all the bindings for “set = 1” etc.

immediate_size: u32

The number of bytes of immediate data that are allocated for use in the shader. The var<immediate>s in the shader attached to this pipeline must be equal or smaller than this size.

If this value is non-zero, wgt::Features::IMMEDIATES must be enabled.