Function wgpu_core::command::bind::compute_nonoverlapping_ranges

pub fn compute_nonoverlapping_ranges(
    ranges: &[PushConstantRange],
) -> ArrayVec<PushConstantRange, { _ }>

Break up possibly overlapping push constant ranges into a set of non-overlapping ranges which contain all the stage flags of the original ranges. This allows us to zero out (or write any value) to every possible value.