wgpu_core::command::clear

Function clear_texture_cmd

pub(super) fn clear_texture_cmd(
    state: &mut EncodingState<'_, '_>,
    dst_texture: Arc<Texture>,
    subresource_range: &ImageSubresourceRange,
) -> Result<(), ClearError>

Validate and encode a “Clear Texture” command.

This function implements CommandEncoder::clear_texture when invoked via the command encoder APIs or trace playback. It has the suffix _cmd to distinguish it from clear_texture. clear_texture, used internally by this function, is a lower-level function that encodes a texture clear operation without validating it.