wgpu_core::command

Module compute

Structs

ComputePass

A pass’s encoder state and its validity are two distinct conditions, i.e., the full matrix of (open, ended) x (valid, invalid) is possible.

ComputePassDescriptor

ComputePassError

Error encountered when performing a compute pass, stored for later reporting when encoding ends.

State 🔒

Enums

ComputePassErrorInner

Error encountered when performing a compute pass.

DispatchError

Functions

dispatch 🔒

dispatch_indirect 🔒

encode_compute_pass 🔒

set_pipeline 🔒

Type Aliases

ArcComputePassDescriptor 🔒

cbindgen:ignore

ComputeBasePass