Module wgpu_core::command::render

Structs

ArcRenderPassColorAttachment ğŸ”’

Describes a color attachment to a render pass.

ArcRenderPassDepthStencilAttachment

Describes a depth/stencil attachment to a render pass.

ArcRenderPassDescriptor ğŸ”’

Describes the attachments of a render pass.

IndexState ğŸ”’

PassChannel

Describes an individual channel within a render pass, such as color, depth, or stencil.

RenderAttachment ğŸ”’

RenderPass

RenderPassColorAttachment

Describes a color attachment to a render pass.

RenderPassDepthStencilAttachment

Describes a depth/stencil attachment to a render pass.

RenderPassDescriptor

Describes the attachments of a render pass.

RenderPassError

Error encountered when performing a render pass.

RenderPassInfo ğŸ”’

State ğŸ”’

VertexBufferState ğŸ”’

VertexState ğŸ”’

Enums

AttachmentErrorLocation

Describes an attachment location in words.

ColorAttachmentError

LoadOp

Operation to perform to the output attachment at the start of a renderpass.

OptionalState ğŸ”’

RenderPassErrorInner

Error encountered when performing a render pass.

StoreOp

Operation to perform to the output attachment at the end of a renderpass.

Constants

MAX_TOTAL_ATTACHMENTS ğŸ”’

Functions

draw ğŸ”’

draw_indexed ğŸ”’

execute_bundle ğŸ”’

insert_debug_marker ğŸ”’

multi_draw_indirect ğŸ”’

multi_draw_indirect_count ğŸ”’

pop_debug_group ğŸ”’

push_debug_group ğŸ”’

set_bind_group ğŸ”’

set_blend_constant ğŸ”’

set_index_buffer ğŸ”’

set_pipeline ğŸ”’

set_push_constant ğŸ”’

set_scissor ğŸ”’

set_stencil_reference ğŸ”’

set_vertex_buffer ğŸ”’

set_viewport ğŸ”’

write_timestamp ğŸ”’

Type Aliases

AttachmentDataVec ğŸ”’