wgpu_core::command

Struct CommandBufferMutable

pub struct CommandBufferMutable {
    pub(crate) encoder: InnerCommandEncoder,
    pub(crate) trackers: Tracker,
    buffer_memory_init_actions: Vec<BufferInitTrackerAction>,
    texture_memory_actions: CommandBufferTextureMemoryActions,
    as_actions: Vec<AsAction>,
    temp_resources: Vec<TempResource>,
    indirect_draw_validation_resources: DrawResources,
    pub(crate) commands: Vec<Command<ArcReferences>>,
}

The mutable state of a CommandBuffer.

Fields
encoder: InnerCommandEncoder

The wgpu_hal::Api::CommandBuffers we’ve built so far, and the encoder they belong to.

trackers: Tracker

All the resources that the commands recorded so far have referred to.

buffer_memory_init_actions: Vec<BufferInitTrackerAction>

The regions of buffers and textures these commands will read and write.

This is used to determine which portions of which buffers/textures we actually need to initialize. If we’re definitely going to write to something before we read from it, we don’t need to clear its contents.

texture_memory_actions: CommandBufferTextureMemoryActionsas_actions: Vec<AsAction>temp_resources: Vec<TempResource>indirect_draw_validation_resources: DrawResourcescommands: Vec<Command<ArcReferences>>
Implementations

impl CommandBufferMutable

pub(crate) fn validate_acceleration_structure_actions( &self, snatch_guard: &SnatchGuard<'_>, command_index_guard: &mut RwLockWriteGuard<'_, CommandIndices>, ) -> Result<(), ValidateAsActionsError>

impl CommandBufferMutable

pub(crate) fn into_baked_commands(self) -> BakedCommands