wgpu_core::command

Struct ComputePass

pub struct ComputePass {
    base: ComputeBasePass,
    parent: Option<Arc<CommandEncoder>>,
    timestamp_writes: Option<PassTimestampWrites<Arc<QuerySet>>>,
    current_bind_groups: BindGroupStateChange<Arc<BindGroup>>,
    current_pipeline: StateChange<Arc<ComputePipeline>>,
}

A pass’s encoder state and its validity are two distinct conditions, i.e., the full matrix of (open, ended) x (valid, invalid) is possible.

The presence or absence of the parent Option indicates the pass’s state. The presence or absence of an error in base.error indicates the pass’s validity.

Fields
base: ComputeBasePass

All pass data & records is stored here.

parent: Option<Arc<CommandEncoder>>

Parent command encoder that this pass records commands into.

If this is Some, then the pass is in WebGPU’s “open” state. If it is None, then the pass is in the “ended” state. See https://www.w3.org/TR/webgpu/#encoder-state

timestamp_writes: Option<PassTimestampWrites<Arc<QuerySet>>>current_bind_groups: BindGroupStateChange<Arc<BindGroup>>current_pipeline: StateChange<Arc<ComputePipeline>>
Implementations

impl ComputePass

fn new( parent: Arc<CommandEncoder>, desc: ComputePassDescriptor<'_, PassTimestampWrites<Arc<QuerySet>>>, ) -> Self

If the parent command encoder is invalid, the returned pass will be invalid.

fn new_invalid( parent: Arc<CommandEncoder>, label: &Label<'_>, err: ComputePassError, ) -> Self

pub fn label(&self) -> Option<&str>

impl ComputePass

pub fn end(&mut self) -> Result<(), EncoderStateError>

impl ComputePass

pub fn set_bind_group( &mut self, index: u32, bind_group: Option<Arc<BindGroup>>, offsets: &[DynamicOffset], ) -> Result<(), PassStateError>

pub fn set_pipeline( &mut self, compute_pipeline: Arc<ComputePipeline>, ) -> Result<(), PassStateError>

pub fn set_immediates( &mut self, offset: u32, data: &[u8], ) -> Result<(), PassStateError>

pub fn dispatch_workgroups( &mut self, groups_x: u32, groups_y: u32, groups_z: u32, ) -> Result<(), PassStateError>

pub fn dispatch_workgroups_indirect( &mut self, buffer: Arc<Buffer>, offset: BufferAddress, ) -> Result<(), PassStateError>

pub fn push_debug_group( &mut self, label: &str, color: u32, ) -> Result<(), PassStateError>

pub fn pop_debug_group(&mut self) -> Result<(), PassStateError>

pub fn insert_debug_marker( &mut self, label: &str, color: u32, ) -> Result<(), PassStateError>

pub fn write_timestamp( &mut self, query_set: Arc<QuerySet>, query_index: u32, ) -> Result<(), PassStateError>

pub fn begin_pipeline_statistics_query( &mut self, query_set: Arc<QuerySet>, query_index: u32, ) -> Result<(), PassStateError>

pub fn end_pipeline_statistics_query(&mut self) -> Result<(), PassStateError>

pub fn transition_resources( &mut self, buffer_transitions: impl Iterator<Item = BufferTransition<Arc<Buffer>>>, texture_transitions: impl Iterator<Item = TextureTransition<Arc<TextureView>>>, ) -> Result<(), PassStateError>