wgpu_core::command
Struct RecordingGuard

pub(crate) struct RecordingGuard<'a> {
    inner: &'a mut CommandEncoderStatus,
}

A guard to enforce error reporting, for a CommandBuffer in the Recording state.

An RecordingGuard holds a mutable reference to a CommandEncoderStatus that has been verified to be in the Recording state. The RecordingGuard dereferences mutably to the CommandBufferMutable that the status holds.

Dropping an RecordingGuard sets the CommandBuffer’s state to CommandEncoderStatus::Error. If your use of the guard was successful, call its mark_successful method to dispose of it.

Fields
inner: &'a mut CommandEncoderStatus
Implementations

impl<'a> RecordingGuard<'a>

pub(crate) fn mark_successful(self)