Module wgpu_core::command::transfer

Enums

CopyError
Error encountered while attempting to do a copy on a command encoder.

CopySide

TransferError
Error encountered while attempting a data transfer.

Constants

BITS_PER_BYTE

Functions

texture_copy_view_to_hal

validate_linear_texture_data
Function copied with minor modifications from webgpu standard https://gpuweb.github.io/gpuweb/#valid-texture-copy-range

validate_texture_copy_range
Function copied with minor modifications from webgpu standard https://gpuweb.github.io/gpuweb/#valid-texture-copy-range

Type Definitions

BufferCopyView

TextureCopyView