wgpu_core::command

Type Alias ArcComputeCommand

pub type ArcComputeCommand = ComputeCommand<ArcReferences>;

cbindgen:ignore

Aliased Type

enum ArcComputeCommand {
    SetBindGroup {
        index: u32,
        num_dynamic_offsets: usize,
        bind_group: Option<Arc<BindGroup>>,
    },
    SetPipeline(Arc<ComputePipeline>),
    SetPushConstant {
        offset: u32,
        size_bytes: u32,
        values_offset: u32,
    },
    Dispatch([u32; 3]),
    DispatchIndirect {
        buffer: Arc<Buffer>,
        offset: u64,
    },
    PushDebugGroup {
        color: u32,
        len: usize,
    },
    PopDebugGroup,
    InsertDebugMarker {
        color: u32,
        len: usize,
    },
    WriteTimestamp {
        query_set: Arc<QuerySet>,
        query_index: u32,
    },
    BeginPipelineStatisticsQuery {
        query_set: Arc<QuerySet>,
        query_index: u32,
    },
    EndPipelineStatisticsQuery,
}

Variants

SetBindGroup

Fields

index: u32

num_dynamic_offsets: usize

bind_group: Option<Arc<BindGroup>>

SetPipeline(Arc<ComputePipeline>)

SetPushConstant

Set a range of push constants to values stored in push_constant_data.

Fields

offset: u32

The byte offset within the push constant storage to write to. This must be a multiple of four.

size_bytes: u32

The number of bytes to write. This must be a multiple of four.

values_offset: u32

Index in push_constant_data of the start of the data to be written.

Note: this is not a byte offset like offset. Rather, it is the index of the first u32 element in push_constant_data to read.

Dispatch([u32; 3])

DispatchIndirect

Fields

buffer: Arc<Buffer>

offset: u64

PushDebugGroup

Fields

color: u32

len: usize

PopDebugGroup

InsertDebugMarker

Fields

color: u32

len: usize

WriteTimestamp

Fields

query_set: Arc<QuerySet>

query_index: u32

BeginPipelineStatisticsQuery

Fields

query_set: Arc<QuerySet>

query_index: u32

EndPipelineStatisticsQuery