wgpu_core::conv

Function check_texture_dimension_size

pub fn check_texture_dimension_size(
    dimension: TextureDimension,
    _: Extent3d,
    sample_size: u32,
    limits: &Limits,
) -> Result<(), TextureDimensionError>

Check the requested texture size against the supported limits.

This function implements the texture size and sample count checks in vtd dimension step. The format checks are elsewhere in create_texture`.

Note that while there is some basic checking of the sample count here, there is an additional set of checks when sample_count > 1 elsewhere in create_texture`.