wgpu_core::device::global

Type Alias MeshPipelineDescriptor

pub type MeshPipelineDescriptor<'a> = MeshPipelineDescriptor<'a, PipelineLayoutId, ShaderModuleId, PipelineCacheId>;

Aliased Type

pub struct MeshPipelineDescriptor<'a> {
    pub label: Option<Cow<'a, str>>,
    pub layout: Option<Id<PipelineLayout>>,
    pub task: Option<TaskState<'a, Id<ShaderModule>>>,
    pub mesh: MeshState<'a, Id<ShaderModule>>,
    pub primitive: PrimitiveState,
    pub depth_stencil: Option<DepthStencilState>,
    pub multisample: MultisampleState,
    pub fragment: Option<FragmentState<'a, Id<ShaderModule>>>,
    pub multiview: Option<NonZero<u32>>,
    pub cache: Option<Id<PipelineCache>>,
}

Fields
label: Option<Cow<'a, str>>layout: Option<Id<PipelineLayout>>

The layout of bind groups for this pipeline.

task: Option<TaskState<'a, Id<ShaderModule>>>

The task processing state for this pipeline.

mesh: MeshState<'a, Id<ShaderModule>>

The mesh processing state for this pipeline

primitive: PrimitiveState

The properties of the pipeline at the primitive assembly and rasterization level.

depth_stencil: Option<DepthStencilState>

The effect of draw calls on the depth and stencil aspects of the output target, if any.

multisample: MultisampleState

The multi-sampling properties of the pipeline.

fragment: Option<FragmentState<'a, Id<ShaderModule>>>

The fragment processing state for this pipeline.

multiview: Option<NonZero<u32>>

If the pipeline will be used with a multiview render pass, this indicates how many array layers the attachments will have.

cache: Option<Id<PipelineCache>>

The pipeline cache to use when creating this pipeline.