wgpu_core::device

Module queue

Structs

EncoderInFlight 🔒

A series of raw CommandBuffers that have been submitted to a queue, and the wgpu_hal::CommandEncoder that built them.

PendingWrites 🔒

A private command encoder for writes made directly on the device or queue.

Queue

Enums

QueueSubmitError

QueueWriteError

TempResource

A texture or buffer to be freed soon.

Functions

validate_command_buffer 🔒

Type Aliases

SubmittedWorkDoneClosuresend_sync