Struct wgpu_core::device::queue::EncoderInFlight

pub(crate) struct EncoderInFlight {
    raw: Box<dyn DynCommandEncoder>,
    cmd_buffers: Vec<Box<dyn DynCommandBuffer>>,
    pub(crate) trackers: Tracker,
    pub(crate) pending_buffers: HashMap<TrackerIndex, Arc<Buffer>, BuildHasherDefault<FxHasher>>,
    pub(crate) pending_textures: HashMap<TrackerIndex, Arc<Texture>, BuildHasherDefault<FxHasher>>,
}

A series of raw CommandBuffers that have been submitted to a queue, and the wgpu_hal::CommandEncoder that built them.

Fields
raw: Box<dyn DynCommandEncoder>cmd_buffers: Vec<Box<dyn DynCommandBuffer>>trackers: Trackerpending_buffers: HashMap<TrackerIndex, Arc<Buffer>, BuildHasherDefault<FxHasher>>

These are the buffers that have been tracked by PendingWrites.

pending_textures: HashMap<TrackerIndex, Arc<Texture>, BuildHasherDefault<FxHasher>>

These are the textures that have been tracked by PendingWrites.

Implementations

impl EncoderInFlight

pub(crate) unsafe fn land(self) -> Box<dyn DynCommandEncoder>

Free all of our command buffers.

Return the command encoder, fully reset and ready to be reused.