Module wgpu_core::hub

Structs

Global

GlobalReport

Hub

HubReport

Hubs

IdentityManager

A simple structure to allocate Id identifiers.

IdentityManagerFactory

Registry

Storage

A table of T values indexed by the id type I.

StorageReport

Enums

Root

Traits

Access

Type system for enforcing the lock order on shared HUB structures. If type A implements Access<B>, that means we are allowed to proceed with locking resource B after we lock A.

GlobalIdentityHandlerFactory

HalApi

IdentityHandler

IdentityHandlerFactory

Resource

Type Definitions

Input