Struct wgpu_core::hub::Storage

pub struct Storage<T, I: TypedId> {
    map: Vec<Element<T>>,
    kind: &'static str,
    _phantom: PhantomData<I>,
}

Fields
map: Vec<Element<T>>kind: &'static str_phantom: PhantomData<I>
Implementations

impl<T, I: TypedId> Storage<T, I>

pub(crate) fn contains(&self, id: I) -> bool

pub(crate) fn get(&self, id: I) -> Result<&T, InvalidId>

Get a reference to an item behind a potentially invalid ID. Panics if there is an epoch mismatch, or the entry is empty.

pub(crate) fn get_mut(&mut self, id: I) -> Result<&mut T, InvalidId>

Get a mutable reference to an item behind a potentially invalid ID. Panics if there is an epoch mismatch, or the entry is empty.

fn insert_impl(&mut self, index: usize, element: Element<T>)

pub(crate) fn insert(&mut self, id: I, value: T)

pub(crate) fn insert_error(&mut self, id: I)

pub(crate) fn remove(&mut self, id: I) -> Option<T>

pub(crate) fn try_remove(&mut self, id: I) -> Option<T>

pub(crate) fn iter(&self, backend: Backend) -> impl Iterator<Item = (I, &T)>