wgpu_core::id

Trait TypedId

pub trait TypedId:
    Copy
    + Debug
    + Any
    + 'static
    + WasmNotSendSync
    + Eq
    + Hash {
    // Required methods
    fn zip(index: u32, epoch: u32, backend: Backend) -> Self;
    fn unzip(self) -> (u32, u32, Backend);
    fn into_raw(self) -> NonZeroU64;
}

Trait carrying methods for direct Id access.

Most wgpu-core clients should not use this trait. Unusual clients that need to construct Id values directly, or access their components, like the WGPU recording player, may use this trait to do so.

Required Methods

fn zip(index: u32, epoch: u32, backend: Backend) -> Self

fn unzip(self) -> (u32, u32, Backend)

fn into_raw(self) -> NonZeroU64

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.