wgpu_core::indirect_validation
Module draw

Structs

BufferPoolEntry ğŸ”’

CurrentEntry ğŸ”’

Draw ğŸ”’

Holds all device-level resources that are needed to validate indirect draws.

DrawBatcher ğŸ”’

Accumulates all needed data needed to validate indirect draws.

DrawIndirectValidationBatch ğŸ”’

DrawResources ğŸ”’

Holds all command buffer-level resources that are needed to validate indirect draws.

MetadataEntry ğŸ”’

This must match the MetadataEntry struct used by the shader.

Constants

BUFFER_SIZE ğŸ”’

Note: This needs to be under:

Functions

calculate_src_buffer_binding_size ğŸ”’

Returns the largest binding size that when combined with dynamic offsets can address the whole buffer.

calculate_src_offsets ğŸ”’

Splits the given offset into a dynamic offset & offset.

create_bind_group_layout ğŸ”’

create_buffer_and_bind_group ğŸ”’

create_validation_module ğŸ”’

create_validation_pipeline ğŸ”’