wgpu_core::lock

Module rank

Ranks for wgpu-core locks, restricting acquisition order.

See LockRank.

Macros

define_lock_ranks 🔒

Define a set of lock ranks, and each rank’s permitted successors.

Structs

LockRank

The rank of a lock.

LockRankSet

A bitflags type representing a set of lock ranks.

Enums

LockRankNumber 🔒

Constants

BLAS_BUILT_INDEX

BLAS_COMPACTION_STATE

BUFFER_BIND_GROUPS

BUFFER_INITIALIZATION_STATUS

BUFFER_MAP_STATE

BUFFER_POOL

COMMAND_ALLOCATOR_FREE_ENCODERS

COMMAND_BUFFER_DATA

DEVICE_COMMAND_INDICES

DEVICE_DEFERRED_DESTROY

DEVICE_FENCE

DEVICE_LOST_CLOSURE

DEVICE_SNATCHABLE_LOCK

DEVICE_TRACE

DEVICE_TRACKERS

DEVICE_USAGE_SCOPES

IDENTITY_MANAGER_VALUES

QUEUE_LIFE_TRACKER

QUEUE_PENDING_WRITES

REGISTRY_STORAGE

RESOURCE_POOL_INNER

SHARED_TRACKER_INDEX_ALLOCATOR_INNER

SURFACE_PRESENTATION

TEXTURE_BIND_GROUPS

TEXTURE_CLEAR_MODE

TEXTURE_INITIALIZATION_STATUS

TEXTURE_VIEWS

TLAS_BUILT_INDEX

TLAS_DEPENDENCIES