wgpu_core::pipeline
Type Alias ResolvedRenderPipelineDescriptor

pub type ResolvedRenderPipelineDescriptor<'a> = RenderPipelineDescriptor<'a, Arc<PipelineLayout>, Arc<ShaderModule>, Arc<PipelineCache>>;

cbindgen:ignore

Aliased Type

struct ResolvedRenderPipelineDescriptor<'a> {
    pub label: Option<Cow<'a, str>>,
    pub layout: Option<Arc<PipelineLayout>>,
    pub vertex: VertexState<'a, Arc<ShaderModule>>,
    pub primitive: PrimitiveState,
    pub depth_stencil: Option<DepthStencilState>,
    pub multisample: MultisampleState,
    pub fragment: Option<FragmentState<'a, Arc<ShaderModule>>>,
    pub multiview: Option<NonZero<u32>>,
    pub cache: Option<Arc<PipelineCache>>,
}

Fields
label: Option<Cow<'a, str>>layout: Option<Arc<PipelineLayout>>

The layout of bind groups for this pipeline.

vertex: VertexState<'a, Arc<ShaderModule>>

The vertex processing state for this pipeline.

primitive: PrimitiveState

The properties of the pipeline at the primitive assembly and rasterization level.

depth_stencil: Option<DepthStencilState>

The effect of draw calls on the depth and stencil aspects of the output target, if any.

multisample: MultisampleState

The multi-sampling properties of the pipeline.

fragment: Option<FragmentState<'a, Arc<ShaderModule>>>

The fragment processing state for this pipeline.

multiview: Option<NonZero<u32>>

If the pipeline will be used with a multiview render pass, this indicates how many array layers the attachments will have.

cache: Option<Arc<PipelineCache>>

The pipeline cache to use when creating this pipeline.