Module wgpu_core::resource

Structs

Buffer

BufferMapCallback

BufferMapCallbackC

BufferMapOperation

QuerySet

Sampler

SamplerDescriptor

Describes a Sampler

StagingBuffer

A temporary buffer, consumed by the command that uses it.

Texture

TextureView

TextureViewDescriptor

Describes a TextureView.

Enums

BufferAccessError

BufferMapAsyncStatus

The status code provided to the buffer mapping callback.

CreateBufferError

CreateQuerySetError

CreateSamplerError

CreateTextureError

CreateTextureViewError

DestroyError

TextureClearMode

TextureDimensionError

TextureErrorDimension

TextureViewDestroyError

Type Definitions

BufferAccessResult

BufferDescriptor

QuerySetDescriptor

TextureDescriptor