wgpu_core::resource

Struct BufferMapping

pub struct BufferMapping {
    ptr: NonNull<u8>,
    len: u64,
    mode: MapMode,
    _guard: RwLockReadGuard<'static, BufferMapState>,
    _lock: Arc<RwLock<BufferMapState>>,
}

Represents a mapped range in the buffer.

While returned BufferMapping is alive, Buffer::destroy and Buffer::unmap calls will block until BufferMapping is dropped.

Fields
ptr: NonNull<u8>len: u64mode: MapMode_guard: RwLockReadGuard<'static, BufferMapState>_lock: Arc<RwLock<BufferMapState>>
Implementations

impl BufferMapping

fn new( ptr: NonNull<u8>, len: u64, mode: MapMode, lock: Arc<RwLock<BufferMapState>>, ) -> Self

pub fn ptr(&self) -> NonNull<u8>

How to safely use the pointer

Pointer is only accessed for the length returned by BufferMapping::len.

Pointer is only accessed while BufferMapping is alive.

Reading is allowed if BufferMapping::mode is [MapMode::Read] (buffer has [wgt::BufferUsages::MAP_READ] set).

Writing is allowed if BufferMapping::mode is [MapMode::Write] (buffer has [wgt::BufferUsages::MAP_WRITE] set).

Atomics must not be used through the pointer since read-modify-write operations may not function on write-combining memory. This also means that it’s not safe to create rust references from this pointer since rust’s memory model requires atomic operations to work.

pub fn mode(&self) -> MapMode

pub fn len(&self) -> u64

pub fn read_into_vec(&self, offset: usize) -> Vec<u8>

Read the contents of the mapped buffer from offset into a Vec<u8>. BufferMapping::mode must be [MapMode::Read] (buffer has [wgt::BufferUsages::MAP_READ] set) otherwise it will panic.

pub fn read(&self, dst: &mut [u8], offset: usize)

Read the contents of the mapped buffer into dst starting from offset. BufferMapping::mode must be [MapMode::Read] (buffer has [wgt::BufferUsages::MAP_READ] set) otherwise it will panic.

pub fn write_slice(&self) -> WriteOnly<'_, [u8]>

Returns a write-only slice to the mapped buffer. BufferMapping::mode must be [MapMode::Write] (buffer has [wgt::BufferUsages::MAP_WRITE] set) otherwise it will panic.