wgpu_core::resource

Struct SamplerDescriptor

pub struct SamplerDescriptor<'a> {
    pub label: Label<'a>,
    pub address_modes: [AddressMode; 3],
    pub mag_filter: FilterMode,
    pub min_filter: FilterMode,
    pub mipmap_filter: FilterMode,
    pub lod_min_clamp: f32,
    pub lod_max_clamp: f32,
    pub compare: Option<CompareFunction>,
    pub anisotropy_clamp: u16,
    pub border_color: Option<SamplerBorderColor>,
}

Describes a Sampler

Fields
label: Label<'a>

Debug label of the sampler.

This will show up in graphics debuggers for easy identification.

address_modes: [AddressMode; 3]

How to deal with out of bounds accesses in the u (i.e. x) direction

mag_filter: FilterMode

How to filter the texture when it needs to be magnified (made larger)

min_filter: FilterMode

How to filter the texture when it needs to be minified (made smaller)

mipmap_filter: FilterMode

How to filter between mip map levels

lod_min_clamp: f32

Minimum level of detail (i.e. mip level) to use

lod_max_clamp: f32

Maximum level of detail (i.e. mip level) to use

compare: Option<CompareFunction>

If this is enabled, this is a comparison sampler using the given comparison function.

anisotropy_clamp: u16

Must be at least 1. If this is not 1, all filter modes must be linear.

border_color: Option<SamplerBorderColor>

Border color to use when address_mode is AddressMode::ClampToBorder