wgpu_core::resource

Struct TextureView

pub struct TextureView {
    pub(crate) raw: Snatchable<Box<dyn DynTextureView>>,
    pub(crate) parent: Arc<Texture>,
    pub(crate) device: Arc<Device>,
    pub(crate) desc: HalTextureViewDescriptor,
    pub(crate) format_features: TextureFormatFeatures,
    pub(crate) render_extent: Result<Extent3d, TextureViewNotRenderableReason>,
    pub(crate) samples: u32,
    pub(crate) selector: TextureSelector,
    pub(crate) label: String,
}

Fields
raw: Snatchable<Box<dyn DynTextureView>>parent: Arc<Texture>device: Arc<Device>desc: HalTextureViewDescriptorformat_features: TextureFormatFeaturesrender_extent: Result<Extent3d, TextureViewNotRenderableReason>

This is Err only if the texture view is not renderable

samples: u32selector: TextureSelectorlabel: String

The label from the descriptor used to create the resource.

Implementations

impl TextureView

fn to_render_attachment(&self, usage: TextureUses) -> RenderAttachment

impl TextureView

pub(crate) fn check_usage( &self, expected: TextureUsages, ) -> Result<(), MissingTextureUsageError>

Checks that the given texture usage contains the required texture usage, returns an error otherwise.