wgpu_core::resource
Trait Labeled

pub trait Labeled: ResourceType {
    // Required method
    fn label(&self) -> &str;

    // Provided method
    fn error_ident(&self) -> ResourceErrorIdent { ... }
}

Required Methods

fn label(&self) -> &str

Returns a string identifying this resource for logging and errors.

It may be a user-provided string or it may be a placeholder from wgpu.

It is non-empty unless the user-provided string was empty.

Provided Methods

fn error_ident(&self) -> ResourceErrorIdent

Object Safety

This trait is not object safe.