wgpu_core::resource

Type Alias ExternalTextureDescriptor

pub type ExternalTextureDescriptor<'a> = ExternalTextureDescriptor<Label<'a>>;

Aliased Type

struct ExternalTextureDescriptor<'a> {
    pub label: Option<Cow<'a, str>>,
    pub width: u32,
    pub height: u32,
    pub format: ExternalTextureFormat,
    pub yuv_conversion_matrix: [f32; 16],
    pub sample_transform: [f32; 6],
    pub load_transform: [f32; 6],
}

Fields
label: Option<Cow<'a, str>>

Debug label of the external texture. This will show up in graphics debuggers for easy identification.

width: u32

Width of the external texture.

height: u32

Height of the external texture.

format: ExternalTextureFormat

Format of the external texture.

yuv_conversion_matrix: [f32; 16]

4x4 column-major matrix with which to convert sampled YCbCr values to RGBA. This is ignored when format is [ExternalTextureFormat::Rgba].

sample_transform: [f32; 6]

Transform to apply to ImageSample coordinates.

This is a 3x2 column-major matrix representing an affine transform from normalized texture coordinates to the normalized coordinates that should be sampled from the external texture’s underlying plane(s).

This transform may scale, translate, flip, and rotate in 90-degree increments, but the result of transforming the rectangle (0,0)..(1,1) must be an axis-aligned rectangle that falls within the bounds of (0,0)..(1,1).

load_transform: [f32; 6]

Transform to apply to ImageLoad coordinates.

This is a 3x2 column-major matrix representing an affine transform from non-normalized texel coordinates to the non-normalized coordinates of the texel that should be loaded from the external texture’s underlying plane 0. For planes 1 and 2, if present, plane 0’s coordinates are scaled according to the textures’ relative sizes.

This transform may scale, translate, flip, and rotate in 90-degree increments, but the result of transforming the rectangle (0,0)..(width, height) must be an axis-aligned rectangle that falls within the bounds of (0,0)..(width, height).