Module wgpu_core::track::buffer

Buffer Trackers

Buffers are represented by a single state for the whole resource,

a 16 bit bitflag of buffer usages. Because there is only ever

one subresource, they have no selector. !

Structs

BufferBindGroupState 🔒

Stores all the buffers that a bind group stores.

BufferTracker 🔒

Stores all buffer state within a command buffer or device.

BufferUsageScope 🔒

Stores all buffer state within a single usage scope.

Enums

BufferStateProvider 🔒

Source of Buffer State.

Functions

barrier 🔒 ⚠

insert 🔒 ⚠

insert_or_barrier_update 🔒 ⚠

If the resource isn’t tracked

insert_or_merge 🔒 ⚠

Does an insertion operation if the index isn’t tracked in the current metadata, otherwise merges the given state with the current state. If the merging would cause a conflict, returns that usage conflict.

merge 🔒 ⚠

update 🔒 ⚠