Function wgpu_core::track::texture::insert_or_merge

unsafe fn insert_or_merge<A: HalApi>(
    texture_data: (&LifeGuard, &TextureSelector),
    current_state_set: &mut TextureStateSet,
    resource_metadata: &mut ResourceMetadata<A>,
    index32: u32,
    index: usize,
    state_provider: TextureStateProvider<'_>,
    metadata_provider: ResourceMetadataProvider<'_, A>
) -> Result<(), UsageConflict>

Does an insertion operation if the index isn’t tracked in the current metadata, otherwise merges the given state with the current state. If the merging would cause a conflict, returns that usage conflict.

Safety

Indexes must be valid indexes into all arrays passed in to this function, either directly or via metadata or provider structs.