wgpu_core::track
Type Alias UsageScopePool

pub(crate) type UsageScopePool = Mutex<Vec<(BufferUsageScope, TextureUsageScope)>>;

A pool for storing the memory used by UsageScopes. We take and store this memory when the scope is dropped to avoid reallocating. The memory required only grows and allocation cost is significant when a large number of resources have been used.

Aliased Type

struct UsageScopePool(Mutex<RawMutex, Vec<(BufferUsageScope, TextureUsageScope)>>);

Fields
0: Mutex<RawMutex, Vec<(BufferUsageScope, TextureUsageScope)>>