Module wgpu_core::validation

Structs

EntryPoint 🔒

Interface

InterfaceVar

MissingBufferUsageError

MissingTextureUsageError

NumericType

Resource 🔒

SpecializationConstant 🔒

Enums

BindingError

FilteringError

InputError

NumericDimension 🔒

ResourceType 🔒

StageError

Errors produced when validating a programmable stage of a pipeline.

Varying 🔒

Functions

check_buffer_usage

Checks that the given buffer usage contains the required buffer usage, returns an error otherwise.

check_texture_format

Return true if the fragment format is covered by the provided output.

check_texture_usage

Checks that the given texture usage contains the required texture usage, returns an error otherwise.

map_storage_format_from_naga 🔒

map_storage_format_to_naga 🔒

Type Aliases

StageIo