Crate wgpu_core

This library safely implements WebGPU on native platforms. It is designed for integration into browsers, as well as wrapping into other language-specific user-friendly libraries.

Feature flags

counters — Internally count resources and events for debugging purposes. If the counters feature is disabled, the counting infrastructure is removed from the build and the exposed counters always return 0.

api_log_info — Log all API entry points at info instead of trace level. Also, promotes certain debug log calls to info.

resource_log_info — Log resource lifecycle management at info instead of trace level.

renderdoc — Support the Renderdoc graphics debugger: https://renderdoc.org/

strict_asserts — Apply run-time checks, even in release builds. These are in addition to the validation carried out at public APIs in all builds.

indirect-validation — Validates indirect draw/dispatch calls. This will also enable naga’s WGSL frontend since we use a WGSL compute shader to do the validation.

serde — Enables serialization via serde on common wgpu types.

trace — Enable API tracing.

observe_locks — Enable lock order observation.

replay — Enable API replaying

raw-window-handle — Enable creating instances using raw-window-handle

wgsl — Enable ShaderModuleSource::Wgsl

glsl — Enable ShaderModuleSource::Glsl

spirv — Enable ShaderModuleSource::SpirV

fragile-send-sync-non-atomic-wasm — Implement Send and Sync on Wasm, but only if atomics are not enabled.

WebGL/WebGPU objects can not be shared between threads. However, it can be useful to artificially mark them as Send and Sync anyways to make it easier to write cross-platform code. This is technically very unsafe in a multithreaded environment, but on a wasm binary compiled without atomics we know we are definitely not in a multithreaded environment.

Backends, passed through to wgpu-hal

metal — Enable the metal backend.

vulkan — Enable the vulkan backend.

gles — Enable the GLES backend.

This is used for all of GLES, OpenGL, and WebGL.

dx12 — Enable the dx12 backend.

Re-exports

pub use validation::map_storage_format_from_naga;

pub use validation::map_storage_format_to_naga;

pub use naga;

Modules

api

binding_model

command

conv 🔒

device

error

global

hal_api

hash_utils 🔒

Module for hashing utilities.

hub

Allocating resource ids, and tracking the resources they refer to.

id

identity

indirect_validation 🔒 indirect-validation

init_tracker 🔒

Lazy initialization of texture and buffer memory.

instance

lock 🔒

Instrumented lock types.

pipeline

pipeline_cache 🔒

pool 🔒

present

Presentation.

ray_tracing

registry

resource

scratch 🔒

snatch 🔒

storage

track 🔒

Resource State and Lifetime Trackers

validation

weak_vec 🔒

Module containing the WeakVec API.

Macros

api_log 🔒 Non-api_log_info

api_log_debug 🔒 Non-api_log_info

impl_labeled

impl_parent_device

impl_resource_type

impl_storage_item

impl_trackable

resource_log 🔒 Non-resource_log_info

Constants

DOWNLEVEL_ERROR_MESSAGE 🔒

DOWNLEVEL_WARNING_MESSAGE 🔒

MAX_BIND_GROUPS

MAX_COLOR_ATTACHMENTS

MAX_VERTEX_BUFFERS

Traits

LabelHelpers 🔒

Functions

get_greatest_common_divisor 🔒

get_lowest_common_denom 🔒

hal_label

Type Aliases

Epoch 🔒

Index 🔒

Label

RawString

SubmissionIndex

The index of a queue submission.