Crate wgpu_core

This library safely implements WebGPU on native platforms. It is designed for integration into browsers, as well as wrapping into other language-specific user-friendly libraries.

Modules

api

binding_model

command

device

error

global

hal_api

hub

Allocating resource ids, and tracking the resources they refer to.

id

identity

instance

pipeline

present

Presentation.

registry

resource

storage

Macros

gfx_if_dx11_hiddenNot (dx11 and Windows)

gfx_if_dx12_hiddenNot (dx12 and Windows)

gfx_if_gles_hiddengles

gfx_if_metal_hiddenNot (metal and (macOS or iOS))

gfx_if_vulkan_hiddenvulkan and non-WebAssembly

gfx_select

Dispatch on an Id’s backend to a backend-generic method.

Structs

LifeGuard

Information needed to decide when it’s safe to free some wgpu-core resource.

Constants

MAX_BIND_GROUPS

MAX_COLOR_ATTACHMENTS

MAX_VERTEX_BUFFERS

Functions

hal_label

Type Aliases

Label

RawString