Crate wgpu_core

This library safely implements WebGPU on native platforms. It is designed for integration into browsers, as well as wrapping into other language-specific user-friendly libraries.

\

Documentation hidden

As a workaround for an issue in rustdoc that affects wgpu-core documentation builds
severely, the documentation for wgpu-core is empty unless built with RUSTFLAGS="--cfg wgpu_core_doc", which may take a very long time.