wgpu_core

Module binding_model

Structs

BindGroup

BindGroupDescriptor

Describes a group of bindings and the resources to be bound.

BindGroupDynamicBindingData

BindGroupEntry

Bindable resource and the slot to bind it to. cbindgen:ignore

BindGroupLayout

Bind group layout.

BindGroupLayoutDescriptor

Describes a BindGroupLayout.

BindingTypeMaxCountError

BindingTypeMaxCountValidator 🔒

BufferBinding

LateMinBufferBindingSizeMismatch

PerStageBindingTypeCounter 🔒

PipelineLayout

PipelineLayoutDescriptor

Describes a pipeline layout.

Enums

BindError

BindGroupLayoutEntryError

BindingError

BindingResource

BindingTypeMaxCountErrorKind

BindingZone

CreateBindGroupError

CreateBindGroupLayoutError

CreatePipelineLayoutError

ExclusivePipeline 🔒

Used by BindGroupLayout. It indicates whether the BGL must be used with a specific pipeline. This constraint only happens when the BGLs have been derived from a pipeline without a layout.

GetBindGroupLayoutError

PushConstantUploadError

Functions

buffer_binding_type_alignment 🔒

buffer_binding_type_bounds_check_alignment 🔒

Type Aliases

ResolvedBindGroupDescriptor

cbindgen:ignore

ResolvedBindGroupEntry

cbindgen:ignore

ResolvedBindingResource

ResolvedBufferBinding

ResolvedPipelineLayoutDescriptor

cbindgen:ignore