Module wgpu_core::binding_model

Structs

BindGroup

BindGroupDescriptor

Describes a group of bindings and the resources to be bound.

BindGroupDynamicBindingData

BindGroupEntry

Bindable resource and the slot to bind it to.

BindGroupLayout

Bind group layout.

BindGroupLayoutDescriptor

Describes a BindGroupLayout.

BindingTypeMaxCountError

BufferBinding

LateMinBufferBindingSizeMismatch

PipelineLayout

PipelineLayoutDescriptor

Describes a pipeline layout.

Enums

BindError

BindGroupLayoutEntryError

BindingResource

BindingTypeMaxCountErrorKind

BindingZone

CreateBindGroupError

CreateBindGroupLayoutError

CreatePipelineLayoutError

GetBindGroupLayoutError

PushConstantUploadError