Module wgpu_core::command

Modules

bundle_ffi

compute_ffi

render_ffi

Structs

BasePassRef

ColorAttachmentDescriptor
Describes a color attachment to a render pass.

CommandBuffer

ComputePass

ComputePassDescriptor

DepthStencilAttachmentDescriptor
Describes a depth/stencil attachment to a render pass.

PassChannel
Describes an individual channel within a render pass, such as color, depth, or stencil.

Rect

RenderBundle

RenderBundleEncoder

RenderBundleEncoderDescriptor
Describes a RenderBundleEncoder.

RenderPass

RenderPassDescriptor
Describes the attachments of a render pass.

Enums

CommandAllocatorError

CommandEncoderError

ComputePassError

CopyError
Error encountered while attempting to do a copy on a command encoder.

CreateRenderBundleError
Error type returned from RenderBundleEncoder::new if the sample count is invalid.

DispatchError

DrawError
Error validating a draw call.

LoadOp
Operation to perform to the output attachment at the start of a renderpass.

RenderBundleError
Error encountered when finishing recording a render bundle.

RenderCommandError
Error encountered when encoding a render command. This is the shared error set between render bundles and passes.

RenderPassError
Error encountered when performing a render pass.

StoreOp
Operation to perform to the output attachment at the end of a renderpass.

TransferError
Error encountered while attempting a data transfer.

UsageConflict

Type Definitions

BufferCopyView

BufferError

RenderBundleDescriptor

TextureCopyView