Module wgpu_core::command

Modules

bundle_ffi

compute_ffi

render_ffi

Structs

BakedCommands

BasePassRef

CommandBuffer

ComputePass

ComputePassDescriptor

ComputePassError

Error encountered when performing a compute pass.

PassChannel

Describes an individual channel within a render pass, such as color, depth, or stencil.

Rect

RenderBundle

RenderBundleEncoder

RenderBundleEncoderDescriptor

Describes a RenderBundleEncoder.

RenderBundleError

Error encountered when finishing recording a render bundle.

RenderPass

RenderPassColorAttachment

Describes a color attachment to a render pass.

RenderPassDepthStencilAttachment

Describes a depth/stencil attachment to a render pass.

RenderPassDescriptor

Describes the attachments of a render pass.

RenderPassError

Error encountered when performing a render pass.

Enums

ClearError

Error encountered while attempting a clear.

ColorAttachmentError

CommandEncoderError

ComputePassErrorInner

Error encountered when performing a compute pass.

CopyError

Error encountered while attempting to do a copy on a command encoder.

CopySide

CreateRenderBundleError

Error type returned from RenderBundleEncoder::new if the sample count is invalid.

DispatchError

DrawError

Error validating a draw call.

ExecutionError

Error type returned from RenderBundleEncoder::new if the sample count is invalid.

LoadOp

Operation to perform to the output attachment at the start of a renderpass.

PassErrorScope

QueryError

Error encountered when dealing with queries

QueryUseError

Error encountered while trying to use queries

RenderCommandError

Error encountered when encoding a render command. This is the shared error set between render bundles and passes.

RenderPassErrorInner

Error encountered when performing a render pass.

ResolveError

Error encountered while trying to resolve a query.

SimplifiedQueryType

StoreOp

Operation to perform to the output attachment at the end of a renderpass.

TransferError

Error encountered while attempting a data transfer.

Type Definitions

ImageCopyBuffer

ImageCopyTexture

ImageCopyTextureTagged

RenderBundleDescriptor