wgpu_core::command

Module bind

Modules

compat 🔒

Structs

Binder 🔒

EntryPayload 🔒

LateBufferBinding 🔒

PushConstantChange 🔒

Enums

BinderError

Functions

compute_nonoverlapping_ranges

Break up possibly overlapping push constant ranges into a set of non-overlapping ranges which contain all the stage flags of the original ranges. This allows us to zero out (or write any value) to every possible value.