Struct wgpu_core::command::CommandBuffer

pub struct CommandBuffer {
    pub(crate) device: Arc<Device>,
    support_clear_texture: bool,
    label: String,
    pub(crate) data: Mutex<Option<CommandBufferMutable>>,
}

A buffer of commands to be submitted to the GPU for execution.

Whereas the WebGPU API uses two separate types for command buffers and encoders, this type is a fusion of the two:

During command recording, this holds a CommandEncoder accepting this buffer’s commands. In this state, the CommandBuffer type behaves like a WebGPU GPUCommandEncoder.

Once command recording is finished by calling Global::command_encoder_finish, no further recording is allowed. The internal CommandEncoder is retained solely as a storage pool for the raw command buffers. In this state, the value behaves like a WebGPU GPUCommandBuffer.

Once a command buffer is submitted to the queue, it is removed from the id registry, and its contents are taken to construct a BakedCommands, whose contents eventually become the property of the submission queue.

Fields
device: Arc<Device>support_clear_texture: boollabel: String

The label from the descriptor used to create the resource.

data: Mutex<Option<CommandBufferMutable>>

The mutable state of this command buffer.

This Option is populated when the command buffer is first created. When this is submitted, dropped, or destroyed, its contents are extracted into a BakedCommands by CommandBufferMutable::into_baked_commands.

Implementations

impl CommandBuffer

pub(crate) fn new( encoder: Box<dyn DynCommandEncoder>, device: &Arc<Device>, label: &Label<'_>, ) -> Self

pub(crate) fn new_invalid(device: &Arc<Device>, label: &Label<'_>) -> Self

pub(crate) fn insert_barriers_from_tracker( raw: &mut dyn DynCommandEncoder, base: &mut Tracker, head: &Tracker, snatch_guard: &SnatchGuard<'_>, )

pub(crate) fn insert_barriers_from_scope( raw: &mut dyn DynCommandEncoder, base: &mut Tracker, head: &UsageScope<'_>, snatch_guard: &SnatchGuard<'_>, )

pub(crate) fn drain_barriers( raw: &mut dyn DynCommandEncoder, base: &mut Tracker, snatch_guard: &SnatchGuard<'_>, )

pub(crate) fn insert_barriers_from_device_tracker( raw: &mut dyn DynCommandEncoder, base: &mut DeviceTracker, head: &Tracker, snatch_guard: &SnatchGuard<'_>, )

impl CommandBuffer

pub fn try_get<'a>( &'a self, ) -> Result<MappedMutexGuard<'a, CommandBufferMutable>, InvalidResourceError>

pub fn try_take<'a>( &'a self, ) -> Result<CommandBufferMutable, InvalidResourceError>