wgpu_core::command
Struct CommandBufferMutable

pub struct CommandBufferMutable {
    pub(crate) encoder: CommandEncoder,
    pub(crate) trackers: Tracker,
    buffer_memory_init_actions: Vec<BufferInitTrackerAction>,
    texture_memory_actions: CommandBufferTextureMemoryActions,
    pub(crate) pending_query_resets: QueryResetMap,
    blas_actions: Vec<BlasAction>,
    tlas_actions: Vec<TlasAction>,
    temp_resources: Vec<TempResource>,
}

The mutable state of a CommandBuffer.

Fields
encoder: CommandEncoder

The wgpu_hal::Api::CommandBuffers we’ve built so far, and the encoder they belong to.

trackers: Tracker

All the resources that the commands recorded so far have referred to.

buffer_memory_init_actions: Vec<BufferInitTrackerAction>

The regions of buffers and textures these commands will read and write.

This is used to determine which portions of which buffers/textures we actually need to initialize. If we’re definitely going to write to something before we read from it, we don’t need to clear its contents.

texture_memory_actions: CommandBufferTextureMemoryActionspending_query_resets: QueryResetMapblas_actions: Vec<BlasAction>tlas_actions: Vec<TlasAction>temp_resources: Vec<TempResource>
Implementations

impl CommandBufferMutable

pub(crate) fn validate_blas_actions( &self, ) -> Result<(), ValidateBlasActionsError>

pub(crate) fn validate_tlas_actions( &self, snatch_guard: &SnatchGuard<'_>, ) -> Result<(), ValidateTlasActionsError>

impl CommandBufferMutable

pub(crate) fn open_encoder_and_tracker( &mut self, ) -> Result<(&mut dyn DynCommandEncoder, &mut Tracker), DeviceError>

pub(crate) fn into_baked_commands(self) -> BakedCommands