wgpu_core::command
Struct ComputePass

pub struct ComputePass {
    base: ComputeBasePass,
    parent: Option<Arc<CommandBuffer>>,
    timestamp_writes: Option<PassTimestampWrites<Arc<QuerySet>>>,
    current_bind_groups: BindGroupStateChange,
    current_pipeline: StateChange<ComputePipelineId>,
}

A pass’s encoder state and its validity are two distinct conditions, i.e., the full matrix of (open, ended) x (valid, invalid) is possible.

The presence or absence of the parent Option indicates the pass’s state. The presence or absence of an error in base.error indicates the pass’s validity.

Fields
base: ComputeBasePass

All pass data & records is stored here.

parent: Option<Arc<CommandBuffer>>

Parent command buffer that this pass records commands into.

If this is Some, then the pass is in WebGPU’s “open” state. If it is None, then the pass is in the “ended” state. See https://www.w3.org/TR/webgpu/#encoder-state

timestamp_writes: Option<PassTimestampWrites<Arc<QuerySet>>>current_bind_groups: BindGroupStateChangecurrent_pipeline: StateChange<ComputePipelineId>
Implementations

impl ComputePass

fn new( parent: Arc<CommandBuffer>, desc: ComputePassDescriptor<'_, PassTimestampWrites<Arc<QuerySet>>>, ) -> Self

If the parent command buffer is invalid, the returned pass will be invalid.

fn new_invalid( parent: Arc<CommandBuffer>, label: &Label<'_>, err: ComputePassError, ) -> Self

pub fn label(&self) -> Option<&str>