wgpu_core::command
Struct RenderBundle

pub struct RenderBundle {
    base: BasePass<ArcRenderCommand>,
    pub(super) is_depth_read_only: bool,
    pub(super) is_stencil_read_only: bool,
    pub(crate) device: Arc<Device>,
    pub(crate) used: RenderBundleScope,
    pub(super) buffer_memory_init_actions: Vec<BufferInitTrackerAction>,
    pub(super) texture_memory_init_actions: Vec<TextureInitTrackerAction>,
    pub(super) context: RenderPassContext,
    label: String,
    pub(crate) tracking_data: TrackingData,
    discard_hal_labels: bool,
}

Fields
base: BasePass<ArcRenderCommand>is_depth_read_only: boolis_stencil_read_only: booldevice: Arc<Device>used: RenderBundleScopebuffer_memory_init_actions: Vec<BufferInitTrackerAction>texture_memory_init_actions: Vec<TextureInitTrackerAction>context: RenderPassContextlabel: String

The label from the descriptor used to create the resource.

tracking_data: TrackingDatadiscard_hal_labels: bool
Implementations

impl RenderBundle

pub(super) unsafe fn execute( &self, raw: &mut dyn DynCommandEncoder, snatch_guard: &SnatchGuard<'_>, ) -> Result<(), ExecutionError>

Actually encode the contents into a native command buffer.

This is partially duplicating the logic of render_pass_end. However the point of this function is to be lighter, since we already had a chance to go through the commands in render_bundle_encoder_finish.

Note that the function isn’t expected to fail, generally. All the validation has already been done by this point. The only failure condition is if some of the used buffers are destroyed.