wgpu_core::command::timestamp_writes

Type Alias ArcPassTimestampWrites

pub type ArcPassTimestampWrites = PassTimestampWrites<Arc<QuerySet>>;

cbindgen:ignore

Aliased Type

struct ArcPassTimestampWrites {
    pub query_set: Arc<QuerySet>,
    pub beginning_of_pass_write_index: Option<u32>,
    pub end_of_pass_write_index: Option<u32>,
}

Fields
query_set: Arc<QuerySet>

The query set to write the timestamps to.

beginning_of_pass_write_index: Option<u32>

The index of the query set at which a start timestamp of this pass is written, if any.

end_of_pass_write_index: Option<u32>

The index of the query set at which an end timestamp of this pass is written, if any.