wgpu_core::command::transfer

Function handle_src_texture_init

fn handle_src_texture_init(
    cmd_buf_data: &mut CommandBufferMutable,
    device: &Device,
    source: &TexelCopyTextureInfo,
    copy_size: &Extent3d,
    texture: &Arc<Texture>,
    snatch_guard: &SnatchGuard<'_>,
) -> Result<(), TransferError>

Prepare a transfer’s source texture.

Ensure the source texture of a transfer is in the right initialization state, and record the state for after the transfer operation.