wgpu_core::command
Type Alias ArcRenderPassColorAttachment

pub type ArcRenderPassColorAttachment = RenderPassColorAttachment<Arc<TextureView>>;

Aliased Type

struct ArcRenderPassColorAttachment {
    pub view: Arc<TextureView>,
    pub resolve_target: Option<Arc<TextureView>>,
    pub load_op: LoadOp<Color>,
    pub store_op: StoreOp,
}

Fields
view: Arc<TextureView>

The view to use as an attachment.

resolve_target: Option<Arc<TextureView>>

The view that will receive the resolved output if multisampling is used.

load_op: LoadOp<Color>

Operation to perform to the output attachment at the start of a renderpass.

This must be clear if it is the first renderpass rendering to a swap chain image.

store_op: StoreOp

Operation to perform to the output attachment at the end of a renderpass.

Implementations

impl ArcRenderPassColorAttachment

fn hal_ops(&self) -> AttachmentOps

fn clear_value(&self) -> Color