Struct wgpu_core::device::life::LifetimeTracker

pub(crate) struct LifetimeTracker<B: Backend> {
    mapped: Vec<Stored<BufferId>>,
    pub future_suspected_buffers: Vec<Stored<BufferId>>,
    pub future_suspected_textures: Vec<Stored<TextureId>>,
    pub suspected_resources: SuspectedResources,
    active: Vec<ActiveSubmission<B>>,
    free_resources: NonReferencedResources<B>,
    ready_to_map: Vec<Valid<BufferId>>,
}

A struct responsible for tracking resource lifetimes.

Here is how host mapping is handled:

When mapping is requested we add the buffer to the life_tracker list of mapped buffers.

When triage_suspected is called, it checks the last submission index associated with each of the mapped buffer, and register the buffer with either a submission in flight, or straight into ready_to_map vector.

When ActiveSubmission is retired, the mapped buffers associated with it are moved to ready_to_map vector.

Finally, handle_mapping issues all the callbacks.

Fields
mapped: Vec<Stored<BufferId>>

Resources that the user has requested be mapped, but are still in use.

future_suspected_buffers: Vec<Stored<BufferId>>

Buffers can be used in a submission that is yet to be made, by the means of write_buffer(), so we have a special place for them.

future_suspected_textures: Vec<Stored<TextureId>>

Textures can be used in the upcoming submission by write_texture.

suspected_resources: SuspectedResources

Resources that are suspected for destruction.

active: Vec<ActiveSubmission<B>>

Resources that are not referenced any more but still used by GPU. Grouped by submissions associated with a fence and a submission index. The active submissions have to be stored in FIFO order: oldest come first.

free_resources: NonReferencedResources<B>

Resources that are neither referenced or used, just life_tracker actual deletion.

ready_to_map: Vec<Valid<BufferId>>
Implementations

impl<B: Backend> LifetimeTracker<B>

pub fn new() -> Self

pub fn track_submission(
&mut self,
index: usize,
fence: B::Fence,
new_suspects: &SuspectedResources,
temp_buffers: impl Iterator<Item = (B::Buffer, MemoryBlock<B>)>
)

pub(crate) fn map(&mut self, value: Valid<BufferId>, ref_count: RefCount)

fn wait_idle(&self, device: &B::Device) -> Result<(), WaitIdleError>

pub fn triage_submissions(
&mut self,
device: &B::Device,
force_wait: bool
) -> Result<usize, WaitIdleError>

Returns the last submission index that is done.

pub fn cleanup(
&mut self,
device: &B::Device,
heaps_mutex: &Mutex<Heaps<B>>,
descriptor_allocator_mutex: &Mutex<DescriptorAllocator<B>>
)

impl<B: GfxBackend> LifetimeTracker<B>

pub(crate) fn triage_suspected<G: GlobalIdentityHandlerFactory>(
&mut self,
hub: &Hub<B, G>,
trackers: &Mutex<TrackerSet>,
trace: Option<&Mutex<Trace>>,
token: &mut Token<Device<B>>
)

pub(crate) fn triage_mapped<G: GlobalIdentityHandlerFactory>(
&mut self,
hub: &Hub<B, G>,
token: &mut Token<Device<B>>
)

pub(crate) fn triage_framebuffers<G: GlobalIdentityHandlerFactory>(
&mut self,
hub: &Hub<B, G>,
framebuffers: &mut HashMap<AttachmentData<Valid<TextureViewId>>, B::Framebuffer, BuildHasherDefault<FxHasher>>,
token: &mut Token<Device<B>>
)

pub(crate) fn handle_mapping<G: GlobalIdentityHandlerFactory>(
&mut self,
hub: &Hub<B, G>,
raw: &B::Device,
trackers: &Mutex<TrackerSet>,
token: &mut Token<Device<B>>
) -> Vec<(BufferMapOperation, BufferMapAsyncStatus)>