Module wgpu_core::device::queue

Structs

EncoderInFlight 🔒

A series of raw CommandBuffers that have been submitted to a queue, and the wgpu_hal::CommandEncoder that built them.

PendingWrites 🔒

A private command encoder for writes made directly on the device or queue.

Queue

SubmittedWorkDoneClosure

SubmittedWorkDoneClosureC

Enums

QueueSubmitError

QueueWriteError

SubmittedWorkDoneClosureInner 🔒

TempResource

A texture or buffer to be freed soon.

Functions

validate_command_buffer 🔒

Type Aliases

SubmittedWorkDoneCallback 🔒