Struct wgpu_core::device::Device

pub struct Device<B: Backend> {

    pub(crate) raw: B::Device,
    pub(crate) adapter_id: Stored<AdapterId>,
    pub(crate) queue_group: QueueGroup<B>,
    pub(crate) cmd_allocator: CommandAllocator<B>,
    mem_allocator: Mutex<Heaps<B>>,
    desc_allocator: Mutex<DescriptorAllocator<B>>,
    pub(crate) life_guard: LifeGuard,
    pub(crate) active_submission_index: usize,
    pub(crate) trackers: Mutex<TrackerSet>,
    pub(crate) render_passes: Mutex<HashMap<AttachmentData<(Attachment, Layout)>, B::RenderPass, BuildHasherDefault<FxHasher>>>,
    pub(crate) framebuffers: Mutex<HashMap<AttachmentData<Valid<TextureViewId>>, B::Framebuffer, BuildHasherDefault<FxHasher>>>,
    life_tracker: Mutex<LifetimeTracker<B>>,
    temp_suspected: SuspectedResources,
    pub(crate) hal_limits: Limits,
    pub(crate) private_features: PrivateFeatures,
    pub(crate) limits: Limits,
    pub(crate) features: Features,
    pending_writes: PendingWrites<B>,
    pub(crate) trace: Option<Mutex<Trace>>,

}

Fields
raw: B::Deviceadapter_id: Stored<AdapterId>queue_group: QueueGroup<B>cmd_allocator: CommandAllocator<B>mem_allocator: Mutex<Heaps<B>>desc_allocator: Mutex<DescriptorAllocator<B>>life_guard: LifeGuardactive_submission_index: usizetrackers: Mutex<TrackerSet>render_passes: Mutex<HashMap<AttachmentData<(Attachment, Layout)>, B::RenderPass, BuildHasherDefault<FxHasher>>>framebuffers: Mutex<HashMap<AttachmentData<Valid<TextureViewId>>, B::Framebuffer, BuildHasherDefault<FxHasher>>>life_tracker: Mutex<LifetimeTracker<B>>temp_suspected: SuspectedResourceshal_limits: Limitsprivate_features: PrivateFeatureslimits: Limitsfeatures: Featurespending_writes: PendingWrites<B>trace: Option<Mutex<Trace>>
Implementations

impl<B: Backend> Device<B>

pub fn borrow_pending_writes(&mut self) -> &mut B::CommandBuffer

fn prepare_stage(
&mut self,
size: BufferAddress
) -> Result<StagingData<B>, DeviceError>

impl<B: GfxBackend> Device<B>

pub(crate) fn new(
raw: B::Device,
adapter_id: Stored<AdapterId>,
queue_group: QueueGroup<B>,
mem_props: MemoryProperties,
hal_limits: Limits,
private_features: PrivateFeatures,
desc: &DeviceDescriptor,
trace_path: Option<&Path>
) -> Result<Self, CreateDeviceError>

pub(crate) fn last_completed_submission_index(&self) -> usize

fn lock_life_internal<'this, 'token: 'this>(
tracker: &'this Mutex<LifetimeTracker<B>>,
_token: &mut Token<'token, Self>
) -> MutexGuard<'this, LifetimeTracker<B>>

pub(crate) fn lock_life<'this, 'token: 'this>(
&'this self,
token: &mut Token<'token, Self>
) -> MutexGuard<'this, LifetimeTracker<B>>

fn maintain<'this, 'token: 'this, G: GlobalIdentityHandlerFactory>(
&'this self,
hub: &Hub<B, G>,
force_wait: bool,
token: &mut Token<'token, Self>
) -> Result<Vec<(BufferMapOperation, BufferMapAsyncStatus)>, WaitIdleError>

fn untrack<'this, 'token: 'this, G: GlobalIdentityHandlerFactory>(
&'this mut self,
hub: &Hub<B, G>,
trackers: &TrackerSet,
token: &mut Token<'token, Self>
)

fn create_buffer(
&self,
self_id: DeviceId,
desc: &BufferDescriptor<'_>,
memory_kind: Kind
) -> Result<Buffer<B>, CreateBufferError>

fn create_texture(
&self,
self_id: DeviceId,
desc: &TextureDescriptor<'_>
) -> Result<Texture<B>, CreateTextureError>

fn create_compatible_render_pass(
&self,
key: &AttachmentData<(Attachment, Layout)>
) -> Result<B::RenderPass, OutOfMemory>

Create a compatible render pass with a given key.

This functions doesn’t consider the following aspects for compatibility:

image layouts

resolve attachments

fn deduplicate_bind_group_layout(
self_id: DeviceId,
entry_map: &HashMap<u32, BindGroupLayoutEntry, BuildHasherDefault<FxHasher>>,
guard: &Storage<BindGroupLayout<B>, BindGroupLayoutId>
) -> Option<BindGroupLayoutId>

fn get_introspection_bind_group_layouts<'a>(
pipeline_layout: &PipelineLayout<B>,
bgl_guard: &'a Storage<BindGroupLayout<B>, BindGroupLayoutId>
) -> IntrospectionBindGroupLayouts<'a>

fn create_bind_group_layout(
&self,
self_id: DeviceId,
label: Option<&str>,
entry_map: HashMap<u32, BindGroupLayoutEntry, BuildHasherDefault<FxHasher>>
) -> Result<BindGroupLayout<B>, CreateBindGroupLayoutError>

fn create_pipeline_layout(
&self,
self_id: DeviceId,
desc: &PipelineLayoutDescriptor<'_>,
bgl_guard: &Storage<BindGroupLayout<B>, BindGroupLayoutId>
) -> Result<PipelineLayout<B>, CreatePipelineLayoutError>

fn wait_for_submit(
&self,
submission_index: usize,
token: &mut Token<'_, Self>
) -> Result<(), WaitIdleError>

impl<B: Backend> Device<B>

pub(crate) fn destroy_bind_group(&self, bind_group: BindGroup<B>)

pub(crate) fn destroy_buffer(&self, buffer: Buffer<B>)

pub(crate) fn destroy_texture(&self, texture: Texture<B>)

pub(crate) fn prepare_to_die(&mut self)

Wait for idle and remove resources that we can, before we die.

pub(crate) fn dispose(self)