Struct wgpu_core::global::Global

pub struct Global {
    pub instance: Instance,
    /* private fields */
}

Fields
instance: Instance
Implementations

impl Global

pub fn command_encoder_clear_buffer( &self, command_encoder_id: CommandEncoderId, dst: BufferId, offset: BufferAddress, size: Option<BufferAddress> ) -> Result<(), ClearError>

pub fn command_encoder_clear_texture( &self, command_encoder_id: CommandEncoderId, dst: TextureId, subresource_range: &ImageSubresourceRange ) -> Result<(), ClearError>

impl Global

pub fn command_encoder_create_compute_pass( &self, encoder_id: CommandEncoderId, desc: &ComputePassDescriptor<'_> ) -> (ComputePass, Option<CommandEncoderError>)

Creates a compute pass.

If creation fails, an invalid pass is returned. Any operation on an invalid pass will return an error.

If successful, puts the encoder into the [CommandEncoderStatus::Locked] state.

pub fn compute_pass_end( &self, pass: &mut ComputePass ) -> Result<(), ComputePassError>

impl Global

pub fn compute_pass_set_bind_group( &self, pass: &mut ComputePass, index: u32, bind_group_id: BindGroupId, offsets: &[DynamicOffset] ) -> Result<(), ComputePassError>

pub fn compute_pass_set_pipeline( &self, pass: &mut ComputePass, pipeline_id: ComputePipelineId ) -> Result<(), ComputePassError>

pub fn compute_pass_set_push_constants( &self, pass: &mut ComputePass, offset: u32, data: &[u8] ) -> Result<(), ComputePassError>

pub fn compute_pass_dispatch_workgroups( &self, pass: &mut ComputePass, groups_x: u32, groups_y: u32, groups_z: u32 ) -> Result<(), ComputePassError>

pub fn compute_pass_dispatch_workgroups_indirect( &self, pass: &mut ComputePass, buffer_id: BufferId, offset: BufferAddress ) -> Result<(), ComputePassError>

pub fn compute_pass_push_debug_group( &self, pass: &mut ComputePass, label: &str, color: u32 ) -> Result<(), ComputePassError>

pub fn compute_pass_pop_debug_group( &self, pass: &mut ComputePass ) -> Result<(), ComputePassError>

pub fn compute_pass_insert_debug_marker( &self, pass: &mut ComputePass, label: &str, color: u32 ) -> Result<(), ComputePassError>

pub fn compute_pass_write_timestamp( &self, pass: &mut ComputePass, query_set_id: QuerySetId, query_index: u32 ) -> Result<(), ComputePassError>

pub fn compute_pass_begin_pipeline_statistics_query( &self, pass: &mut ComputePass, query_set_id: QuerySetId, query_index: u32 ) -> Result<(), ComputePassError>

pub fn compute_pass_end_pipeline_statistics_query( &self, pass: &mut ComputePass ) -> Result<(), ComputePassError>

impl Global

pub fn command_encoder_write_timestamp( &self, command_encoder_id: CommandEncoderId, query_set_id: QuerySetId, query_index: u32 ) -> Result<(), QueryError>

pub fn command_encoder_resolve_query_set( &self, command_encoder_id: CommandEncoderId, query_set_id: QuerySetId, start_query: u32, query_count: u32, destination: BufferId, destination_offset: BufferAddress ) -> Result<(), QueryError>

impl Global

pub fn command_encoder_create_render_pass( &self, encoder_id: CommandEncoderId, desc: &RenderPassDescriptor<'_> ) -> (RenderPass, Option<CommandEncoderError>)

Creates a render pass.

If creation fails, an invalid pass is returned. Any operation on an invalid pass will return an error.

If successful, puts the encoder into the [CommandEncoderStatus::Locked] state.

impl Global

pub fn render_pass_set_bind_group( &self, pass: &mut RenderPass, index: u32, bind_group_id: BindGroupId, offsets: &[DynamicOffset] ) -> Result<(), RenderPassError>

pub fn render_pass_set_pipeline( &self, pass: &mut RenderPass, pipeline_id: RenderPipelineId ) -> Result<(), RenderPassError>

pub fn render_pass_set_index_buffer( &self, pass: &mut RenderPass, buffer_id: BufferId, index_format: IndexFormat, offset: BufferAddress, size: Option<BufferSize> ) -> Result<(), RenderPassError>

pub fn render_pass_set_vertex_buffer( &self, pass: &mut RenderPass, slot: u32, buffer_id: BufferId, offset: BufferAddress, size: Option<BufferSize> ) -> Result<(), RenderPassError>

pub fn render_pass_set_blend_constant( &self, pass: &mut RenderPass, color: Color ) -> Result<(), RenderPassError>

pub fn render_pass_set_stencil_reference( &self, pass: &mut RenderPass, value: u32 ) -> Result<(), RenderPassError>

pub fn render_pass_set_viewport( &self, pass: &mut RenderPass, x: f32, y: f32, w: f32, h: f32, depth_min: f32, depth_max: f32 ) -> Result<(), RenderPassError>

pub fn render_pass_set_scissor_rect( &self, pass: &mut RenderPass, x: u32, y: u32, w: u32, h: u32 ) -> Result<(), RenderPassError>

pub fn render_pass_set_push_constants( &self, pass: &mut RenderPass, stages: ShaderStages, offset: u32, data: &[u8] ) -> Result<(), RenderPassError>

pub fn render_pass_draw( &self, pass: &mut RenderPass, vertex_count: u32, instance_count: u32, first_vertex: u32, first_instance: u32 ) -> Result<(), RenderPassError>

pub fn render_pass_draw_indexed( &self, pass: &mut RenderPass, index_count: u32, instance_count: u32, first_index: u32, base_vertex: i32, first_instance: u32 ) -> Result<(), RenderPassError>

pub fn render_pass_draw_indirect( &self, pass: &mut RenderPass, buffer_id: BufferId, offset: BufferAddress ) -> Result<(), RenderPassError>

pub fn render_pass_draw_indexed_indirect( &self, pass: &mut RenderPass, buffer_id: BufferId, offset: BufferAddress ) -> Result<(), RenderPassError>

pub fn render_pass_multi_draw_indirect( &self, pass: &mut RenderPass, buffer_id: BufferId, offset: BufferAddress, count: u32 ) -> Result<(), RenderPassError>

pub fn render_pass_multi_draw_indexed_indirect( &self, pass: &mut RenderPass, buffer_id: BufferId, offset: BufferAddress, count: u32 ) -> Result<(), RenderPassError>

pub fn render_pass_multi_draw_indirect_count( &self, pass: &mut RenderPass, buffer_id: BufferId, offset: BufferAddress, count_buffer_id: BufferId, count_buffer_offset: BufferAddress, max_count: u32 ) -> Result<(), RenderPassError>

pub fn render_pass_multi_draw_indexed_indirect_count( &self, pass: &mut RenderPass, buffer_id: BufferId, offset: BufferAddress, count_buffer_id: BufferId, count_buffer_offset: BufferAddress, max_count: u32 ) -> Result<(), RenderPassError>

pub fn render_pass_push_debug_group( &self, pass: &mut RenderPass, label: &str, color: u32 ) -> Result<(), RenderPassError>

pub fn render_pass_pop_debug_group( &self, pass: &mut RenderPass ) -> Result<(), RenderPassError>

pub fn render_pass_insert_debug_marker( &self, pass: &mut RenderPass, label: &str, color: u32 ) -> Result<(), RenderPassError>

pub fn render_pass_write_timestamp( &self, pass: &mut RenderPass, query_set_id: QuerySetId, query_index: u32 ) -> Result<(), RenderPassError>

pub fn render_pass_begin_occlusion_query( &self, pass: &mut RenderPass, query_index: u32 ) -> Result<(), RenderPassError>

pub fn render_pass_end_occlusion_query( &self, pass: &mut RenderPass ) -> Result<(), RenderPassError>

pub fn render_pass_begin_pipeline_statistics_query( &self, pass: &mut RenderPass, query_set_id: QuerySetId, query_index: u32 ) -> Result<(), RenderPassError>

pub fn render_pass_end_pipeline_statistics_query( &self, pass: &mut RenderPass ) -> Result<(), RenderPassError>

pub fn render_pass_execute_bundles( &self, pass: &mut RenderPass, render_bundle_ids: &[RenderBundleId] ) -> Result<(), RenderPassError>

impl Global

pub fn command_encoder_copy_buffer_to_buffer( &self, command_encoder_id: CommandEncoderId, source: BufferId, source_offset: BufferAddress, destination: BufferId, destination_offset: BufferAddress, size: BufferAddress ) -> Result<(), CopyError>

pub fn command_encoder_copy_buffer_to_texture( &self, command_encoder_id: CommandEncoderId, source: &ImageCopyBuffer, destination: &ImageCopyTexture, copy_size: &Extent3d ) -> Result<(), CopyError>

pub fn command_encoder_copy_texture_to_buffer( &self, command_encoder_id: CommandEncoderId, source: &ImageCopyTexture, destination: &ImageCopyBuffer, copy_size: &Extent3d ) -> Result<(), CopyError>

pub fn command_encoder_copy_texture_to_texture( &self, command_encoder_id: CommandEncoderId, source: &ImageCopyTexture, destination: &ImageCopyTexture, copy_size: &Extent3d ) -> Result<(), CopyError>

impl Global

pub fn command_encoder_finish( &self, encoder_id: CommandEncoderId, _desc: &CommandBufferDescriptor<Label<'_>> ) -> (CommandBufferId, Option<CommandEncoderError>)

pub fn command_encoder_push_debug_group( &self, encoder_id: CommandEncoderId, label: &str ) -> Result<(), CommandEncoderError>

pub fn command_encoder_insert_debug_marker( &self, encoder_id: CommandEncoderId, label: &str ) -> Result<(), CommandEncoderError>

pub fn command_encoder_pop_debug_group( &self, encoder_id: CommandEncoderId ) -> Result<(), CommandEncoderError>

impl Global

pub fn adapter_is_surface_supported( &self, adapter_id: AdapterId, surface_id: SurfaceId ) -> Result<bool, IsSurfaceSupportedError>

pub fn surface_get_capabilities( &self, surface_id: SurfaceId, adapter_id: AdapterId ) -> Result<SurfaceCapabilities, GetSurfaceSupportError>

pub fn device_features( &self, device_id: DeviceId ) -> Result<Features, DeviceError>

pub fn device_limits(&self, device_id: DeviceId) -> Result<Limits, DeviceError>

pub fn device_downlevel_properties( &self, device_id: DeviceId ) -> Result<DownlevelCapabilities, DeviceError>

pub fn device_create_buffer( &self, device_id: DeviceId, desc: &BufferDescriptor<'_>, id_in: Option<BufferId> ) -> (BufferId, Option<CreateBufferError>)

pub fn create_buffer_error(&self, backend: Backend, id_in: Option<BufferId>)

Assign id_in an error with the given label.

Ensure that future attempts to use id_in as a buffer ID will propagate the error, following the WebGPU “contagious invalidity” style.

Firefox uses this function to comply strictly with the WebGPU spec, which requires GPUBufferDescriptor validation to be generated on the Device timeline and leave the newly created GPUBuffer invalid.

Ideally, we would simply let device_create_buffer take care of all of this, but some errors must be detected before we can even construct a wgpu_types::BufferDescriptor to give it. For example, the WebGPU API allows a GPUBufferDescriptor’s usage property to be any WebIDL unsigned long value, but we can’t construct a wgpu_types::BufferUsages value from values with unassigned bits set. This means we must validate usage before we can call device_create_buffer.

When that validation fails, we must arrange for the buffer id to be considered invalid. This method provides the means to do so.

pub fn create_render_bundle_error( &self, backend: Backend, id_in: Option<RenderBundleId> )

pub fn create_texture_error(&self, backend: Backend, id_in: Option<TextureId>)

Assign id_in an error with the given label.

See create_buffer_error for more context and explanation.

pub fn device_set_buffer_data( &self, buffer_id: BufferId, offset: BufferAddress, data: &[u8] ) -> BufferAccessResult

Available on crate feature replay only.

pub fn buffer_destroy(&self, buffer_id: BufferId) -> Result<(), DestroyError>

pub fn buffer_drop(&self, buffer_id: BufferId)

pub fn device_create_texture( &self, device_id: DeviceId, desc: &TextureDescriptor<'_>, id_in: Option<TextureId> ) -> (TextureId, Option<CreateTextureError>)

pub unsafe fn create_texture_from_hal( &self, hal_texture: Box<dyn DynTexture>, device_id: DeviceId, desc: &TextureDescriptor<'_>, id_in: Option<TextureId> ) -> (TextureId, Option<CreateTextureError>)

Safety

hal_texture must be created from device_id corresponding raw handle.

hal_texture must be created respecting desc

hal_texture must be initialized

pub unsafe fn create_buffer_from_hal<A: HalApi>( &self, hal_buffer: A::Buffer, device_id: DeviceId, desc: &BufferDescriptor<'_>, id_in: Option<BufferId> ) -> (BufferId, Option<CreateBufferError>)

Safety

hal_buffer must be created from device_id corresponding raw handle.

hal_buffer must be created respecting desc

hal_buffer must be initialized

pub fn texture_destroy(&self, texture_id: TextureId) -> Result<(), DestroyError>

pub fn texture_drop(&self, texture_id: TextureId)

pub fn texture_create_view( &self, texture_id: TextureId, desc: &TextureViewDescriptor<'_>, id_in: Option<TextureViewId> ) -> (TextureViewId, Option<CreateTextureViewError>)

pub fn texture_view_drop( &self, texture_view_id: TextureViewId ) -> Result<(), TextureViewDestroyError>

pub fn device_create_sampler( &self, device_id: DeviceId, desc: &SamplerDescriptor<'_>, id_in: Option<SamplerId> ) -> (SamplerId, Option<CreateSamplerError>)

pub fn sampler_drop(&self, sampler_id: SamplerId)

pub fn device_create_bind_group_layout( &self, device_id: DeviceId, desc: &BindGroupLayoutDescriptor<'_>, id_in: Option<BindGroupLayoutId> ) -> (BindGroupLayoutId, Option<CreateBindGroupLayoutError>)

pub fn bind_group_layout_drop(&self, bind_group_layout_id: BindGroupLayoutId)

pub fn device_create_pipeline_layout( &self, device_id: DeviceId, desc: &PipelineLayoutDescriptor<'_>, id_in: Option<PipelineLayoutId> ) -> (PipelineLayoutId, Option<CreatePipelineLayoutError>)

pub fn pipeline_layout_drop(&self, pipeline_layout_id: PipelineLayoutId)

pub fn device_create_bind_group( &self, device_id: DeviceId, desc: &BindGroupDescriptor<'_>, id_in: Option<BindGroupId> ) -> (BindGroupId, Option<CreateBindGroupError>)

pub fn bind_group_drop(&self, bind_group_id: BindGroupId)

pub fn device_create_shader_module( &self, device_id: DeviceId, desc: &ShaderModuleDescriptor<'_>, source: ShaderModuleSource<'_>, id_in: Option<ShaderModuleId> ) -> (ShaderModuleId, Option<CreateShaderModuleError>)

Create a shader module with the given source.

This function may consume a lot of stack space. Compiler-enforced limits for parsing recursion exist; if shader compilation runs into them, it will return an error gracefully. However, on some build profiles and platforms, the default stack size for a thread may be exceeded before this limit is reached during parsing. Callers should ensure that there is enough stack space for this, particularly if calls to this method are exposed to user input.

pub unsafe fn device_create_shader_module_spirv( &self, device_id: DeviceId, desc: &ShaderModuleDescriptor<'_>, source: Cow<'_, [u32]>, id_in: Option<ShaderModuleId> ) -> (ShaderModuleId, Option<CreateShaderModuleError>)

Safety

This function passes SPIR-V binary to the backend as-is and can potentially result in a driver crash.

pub fn shader_module_drop(&self, shader_module_id: ShaderModuleId)

pub fn device_create_command_encoder( &self, device_id: DeviceId, desc: &CommandEncoderDescriptor<Label<'_>>, id_in: Option<CommandEncoderId> ) -> (CommandEncoderId, Option<DeviceError>)

pub fn command_encoder_drop(&self, command_encoder_id: CommandEncoderId)

pub fn command_buffer_drop(&self, command_buffer_id: CommandBufferId)

pub fn device_create_render_bundle_encoder( &self, device_id: DeviceId, desc: &RenderBundleEncoderDescriptor<'_> ) -> (*mut RenderBundleEncoder, Option<CreateRenderBundleError>)

pub fn render_bundle_encoder_finish( &self, bundle_encoder: RenderBundleEncoder, desc: &RenderBundleDescriptor<'_>, id_in: Option<RenderBundleId> ) -> (RenderBundleId, Option<RenderBundleError>)

pub fn render_bundle_drop(&self, render_bundle_id: RenderBundleId)

pub fn device_create_query_set( &self, device_id: DeviceId, desc: &QuerySetDescriptor<'_>, id_in: Option<QuerySetId> ) -> (QuerySetId, Option<CreateQuerySetError>)

pub fn query_set_drop(&self, query_set_id: QuerySetId)

pub fn device_create_render_pipeline( &self, device_id: DeviceId, desc: &RenderPipelineDescriptor<'_>, id_in: Option<RenderPipelineId>, implicit_pipeline_ids: Option<ImplicitPipelineIds<'_>> ) -> (RenderPipelineId, Option<CreateRenderPipelineError>)

pub fn render_pipeline_get_bind_group_layout( &self, pipeline_id: RenderPipelineId, index: u32, id_in: Option<BindGroupLayoutId> ) -> (BindGroupLayoutId, Option<GetBindGroupLayoutError>)

Get an ID of one of the bind group layouts. The ID adds a refcount, which needs to be released by calling bind_group_layout_drop.

pub fn render_pipeline_drop(&self, render_pipeline_id: RenderPipelineId)

pub fn device_create_compute_pipeline( &self, device_id: DeviceId, desc: &ComputePipelineDescriptor<'_>, id_in: Option<ComputePipelineId>, implicit_pipeline_ids: Option<ImplicitPipelineIds<'_>> ) -> (ComputePipelineId, Option<CreateComputePipelineError>)

pub fn compute_pipeline_get_bind_group_layout( &self, pipeline_id: ComputePipelineId, index: u32, id_in: Option<BindGroupLayoutId> ) -> (BindGroupLayoutId, Option<GetBindGroupLayoutError>)

Get an ID of one of the bind group layouts. The ID adds a refcount, which needs to be released by calling bind_group_layout_drop.

pub fn compute_pipeline_drop(&self, compute_pipeline_id: ComputePipelineId)

pub unsafe fn device_create_pipeline_cache( &self, device_id: DeviceId, desc: &PipelineCacheDescriptor<'_>, id_in: Option<PipelineCacheId> ) -> (PipelineCacheId, Option<CreatePipelineCacheError>)

Safety

The data argument of desc must have been returned by Self::pipeline_cache_get_data for the same adapter

pub fn pipeline_cache_drop(&self, pipeline_cache_id: PipelineCacheId)

pub fn surface_configure( &self, surface_id: SurfaceId, device_id: DeviceId, config: &SurfaceConfiguration<Vec<TextureFormat>> ) -> Option<ConfigureSurfaceError>

pub fn device_poll( &self, device_id: DeviceId, maintain: Maintain<SubmissionIndex> ) -> Result<bool, WaitIdleError>

Check device_id for freeable resources and completed buffer mappings.

Return queue_empty indicating whether there are more queue submissions still in flight.

pub fn poll_all_devices(&self, force_wait: bool) -> Result<bool, WaitIdleError>

Poll all devices on all backends.

This is the implementation of wgpu::Instance::poll_all.

Return all_queue_empty indicating whether there are more queue submissions still in flight.

pub fn device_start_capture(&self, id: DeviceId)

pub fn device_stop_capture(&self, id: DeviceId)

pub fn device_make_invalid(&self, device_id: DeviceId)

pub fn pipeline_cache_get_data(&self, id: PipelineCacheId) -> Option<Vec<u8>>

pub fn device_drop(&self, device_id: DeviceId)

pub fn device_set_device_lost_closure( &self, device_id: DeviceId, device_lost_closure: DeviceLostClosure )

pub fn device_destroy(&self, device_id: DeviceId)

pub fn device_mark_lost(&self, device_id: DeviceId, message: &str)

pub fn device_get_internal_counters( &self, device_id: DeviceId ) -> InternalCounters

pub fn device_generate_allocator_report( &self, device_id: DeviceId ) -> Option<AllocatorReport>

pub fn queue_drop(&self, queue_id: QueueId)

pub fn buffer_map_async( &self, buffer_id: BufferId, offset: BufferAddress, size: Option<BufferAddress>, op: BufferMapOperation ) -> BufferAccessResult

pub fn buffer_get_mapped_range( &self, buffer_id: BufferId, offset: BufferAddress, size: Option<BufferAddress> ) -> Result<(NonNull<u8>, u64), BufferAccessError>

pub fn buffer_unmap(&self, buffer_id: BufferId) -> BufferAccessResult

impl Global

pub fn queue_write_buffer( &self, queue_id: QueueId, buffer_id: BufferId, buffer_offset: BufferAddress, data: &[u8] ) -> Result<(), QueueWriteError>

pub fn queue_create_staging_buffer( &self, queue_id: QueueId, buffer_size: BufferSize, id_in: Option<StagingBufferId> ) -> Result<(StagingBufferId, NonNull<u8>), QueueWriteError>

pub fn queue_write_staging_buffer( &self, queue_id: QueueId, buffer_id: BufferId, buffer_offset: BufferAddress, staging_buffer_id: StagingBufferId ) -> Result<(), QueueWriteError>

pub fn queue_validate_write_buffer( &self, _queue_id: QueueId, buffer_id: BufferId, buffer_offset: u64, buffer_size: BufferSize ) -> Result<(), QueueWriteError>

pub fn queue_write_texture( &self, queue_id: QueueId, destination: &ImageCopyTexture, data: &[u8], data_layout: &ImageDataLayout, size: &Extent3d ) -> Result<(), QueueWriteError>

pub fn queue_submit( &self, queue_id: QueueId, command_buffer_ids: &[CommandBufferId] ) -> Result<SubmissionIndex, QueueSubmitError>

pub fn queue_get_timestamp_period( &self, queue_id: QueueId ) -> Result<f32, InvalidQueue>

pub fn queue_on_submitted_work_done( &self, queue_id: QueueId, closure: SubmittedWorkDoneClosure ) -> Result<(), InvalidQueue>

impl Global

pub fn new(name: &str, instance_desc: InstanceDescriptor) -> Self

pub unsafe fn from_hal_instance<A: HalApi>( name: &str, hal_instance: A::Instance ) -> Self

Safety

Refer to the creation of wgpu-hal Instance for every backend.

pub unsafe fn instance_as_hal<A: HalApi>(&self) -> Option<&A::Instance>

Safety

The raw instance handle returned must not be manually destroyed.

pub unsafe fn from_instance(instance: Instance) -> Self

Safety

The raw handles obtained from the Instance must not be manually destroyed

pub fn generate_report(&self) -> GlobalReport

impl Global

pub unsafe fn instance_create_surface( &self, display_handle: RawDisplayHandle, window_handle: RawWindowHandle, id_in: Option<SurfaceId> ) -> Result<SurfaceId, CreateSurfaceError>

Available on crate feature raw-window-handle only.

Creates a new surface targeting the given display/window handles.

Internally attempts to create hal surfaces for all enabled backends.

Fails only if creation for surfaces for all enabled backends fails in which case the error for each enabled backend is listed. Vice versa, if creation for any backend succeeds, success is returned. Surface creation errors are logged to the debug log in any case.

id_in:

If Some, the id to assign to the surface. A new one will be generated otherwise.

Safety

display_handle must be a valid object to create a surface upon.

window_handle must remain valid as long as the returned SurfaceId is being used.

pub fn surface_drop(&self, id: SurfaceId)

pub fn enumerate_adapters( &self, inputs: AdapterInputs<'_, Adapter> ) -> Vec<AdapterId>

pub fn request_adapter( &self, desc: &RequestAdapterOptions, inputs: AdapterInputs<'_, Adapter> ) -> Result<AdapterId, RequestAdapterError>

pub unsafe fn create_adapter_from_hal( &self, hal_adapter: DynExposedAdapter, input: Option<AdapterId> ) -> AdapterId

Safety

hal_adapter must be created from this global internal instance handle.

pub fn adapter_get_info( &self, adapter_id: AdapterId ) -> Result<AdapterInfo, InvalidAdapter>

pub fn adapter_get_texture_format_features( &self, adapter_id: AdapterId, format: TextureFormat ) -> Result<TextureFormatFeatures, InvalidAdapter>

pub fn adapter_features( &self, adapter_id: AdapterId ) -> Result<Features, InvalidAdapter>

pub fn adapter_limits( &self, adapter_id: AdapterId ) -> Result<Limits, InvalidAdapter>

pub fn adapter_downlevel_capabilities( &self, adapter_id: AdapterId ) -> Result<DownlevelCapabilities, InvalidAdapter>

pub fn adapter_get_presentation_timestamp( &self, adapter_id: AdapterId ) -> Result<PresentationTimestamp, InvalidAdapter>

pub fn adapter_drop(&self, adapter_id: AdapterId)

impl Global

pub fn adapter_request_device( &self, adapter_id: AdapterId, desc: &DeviceDescriptor<'_>, trace_path: Option<&Path>, device_id_in: Option<DeviceId>, queue_id_in: Option<QueueId> ) -> (DeviceId, QueueId, Option<RequestDeviceError>)

pub unsafe fn create_device_from_hal( &self, adapter_id: AdapterId, hal_device: DynOpenDevice, desc: &DeviceDescriptor<'_>, trace_path: Option<&Path>, device_id_in: Option<DeviceId>, queue_id_in: Option<QueueId> ) -> (DeviceId, QueueId, Option<RequestDeviceError>)

Safety

hal_device must be created from adapter_id or its internal handle.

desc must be a subset of hal_device features and limits.

impl Global

pub fn surface_get_current_texture( &self, surface_id: SurfaceId, texture_id_in: Option<TextureId> ) -> Result<SurfaceOutput, SurfaceError>

pub fn surface_present( &self, surface_id: SurfaceId ) -> Result<Status, SurfaceError>

pub fn surface_texture_discard( &self, surface_id: SurfaceId ) -> Result<(), SurfaceError>

impl Global

pub unsafe fn buffer_as_hal<A: HalApi, F: FnOnce(Option<&A::Buffer>) -> R, R>( &self, id: BufferId, hal_buffer_callback: F ) -> R

Safety

The raw buffer handle must not be manually destroyed

pub unsafe fn texture_as_hal<A: HalApi, F: FnOnce(Option<&A::Texture>) -> R, R>( &self, id: TextureId, hal_texture_callback: F ) -> R

Safety

The raw texture handle must not be manually destroyed

pub unsafe fn texture_view_as_hal<A: HalApi, F: FnOnce(Option<&A::TextureView>) -> R, R>( &self, id: TextureViewId, hal_texture_view_callback: F ) -> R

Safety

The raw texture view handle must not be manually destroyed

pub unsafe fn adapter_as_hal<A: HalApi, F: FnOnce(Option<&A::Adapter>) -> R, R>( &self, id: AdapterId, hal_adapter_callback: F ) -> R

Safety

The raw adapter handle must not be manually destroyed

pub unsafe fn device_as_hal<A: HalApi, F: FnOnce(Option<&A::Device>) -> R, R>( &self, id: DeviceId, hal_device_callback: F ) -> R

Safety

The raw device handle must not be manually destroyed

pub unsafe fn device_fence_as_hal<A: HalApi, F: FnOnce(Option<&A::Fence>) -> R, R>( &self, id: DeviceId, hal_fence_callback: F ) -> R

Safety

The raw fence handle must not be manually destroyed

pub unsafe fn surface_as_hal<A: HalApi, F: FnOnce(Option<&A::Surface>) -> R, R>( &self, id: SurfaceId, hal_surface_callback: F ) -> R

Safety

The raw surface handle must not be manually destroyed

pub unsafe fn command_encoder_as_hal_mut<A: HalApi, F: FnOnce(Option<&mut A::CommandEncoder>) -> R, R>( &self, id: CommandEncoderId, hal_command_encoder_callback: F ) -> R

Safety

The raw command encoder handle must not be manually destroyed