Struct wgpu_core::global::Global

pub struct Global {
    pub instance: Instance,
    pub surfaces: Registry<Surface>,
    /* private fields */
}

Fields
instance: Instancesurfaces: Registry<Surface>
Implementations

impl Global

pub fn command_encoder_clear_buffer<A: HalApi>( &self, command_encoder_id: CommandEncoderId, dst: BufferId, offset: BufferAddress, size: Option<BufferAddress> ) -> Result<(), ClearError>

pub fn command_encoder_clear_texture<A: HalApi>( &self, command_encoder_id: CommandEncoderId, dst: TextureId, subresource_range: &ImageSubresourceRange ) -> Result<(), ClearError>

impl Global

pub fn command_encoder_run_compute_pass<A: HalApi>( &self, encoder_id: CommandEncoderId, pass: &ComputePass ) -> Result<(), ComputePassError>

impl Global

pub fn command_encoder_write_timestamp<A: HalApi>( &self, command_encoder_id: CommandEncoderId, query_set_id: QuerySetId, query_index: u32 ) -> Result<(), QueryError>

pub fn command_encoder_resolve_query_set<A: HalApi>( &self, command_encoder_id: CommandEncoderId, query_set_id: QuerySetId, start_query: u32, query_count: u32, destination: BufferId, destination_offset: BufferAddress ) -> Result<(), QueryError>

impl Global

pub fn command_encoder_run_render_pass<A: HalApi>( &self, encoder_id: CommandEncoderId, pass: &RenderPass ) -> Result<(), RenderPassError>

impl Global

pub fn command_encoder_copy_buffer_to_buffer<A: HalApi>( &self, command_encoder_id: CommandEncoderId, source: BufferId, source_offset: BufferAddress, destination: BufferId, destination_offset: BufferAddress, size: BufferAddress ) -> Result<(), CopyError>

pub fn command_encoder_copy_buffer_to_texture<A: HalApi>( &self, command_encoder_id: CommandEncoderId, source: &ImageCopyBuffer, destination: &ImageCopyTexture, copy_size: &Extent3d ) -> Result<(), CopyError>

pub fn command_encoder_copy_texture_to_buffer<A: HalApi>( &self, command_encoder_id: CommandEncoderId, source: &ImageCopyTexture, destination: &ImageCopyBuffer, copy_size: &Extent3d ) -> Result<(), CopyError>

pub fn command_encoder_copy_texture_to_texture<A: HalApi>( &self, command_encoder_id: CommandEncoderId, source: &ImageCopyTexture, destination: &ImageCopyTexture, copy_size: &Extent3d ) -> Result<(), CopyError>

impl Global

pub fn command_encoder_finish<A: HalApi>( &self, encoder_id: CommandEncoderId, _desc: &CommandBufferDescriptor<Label<'_>> ) -> (CommandBufferId, Option<CommandEncoderError>)

pub fn command_encoder_push_debug_group<A: HalApi>( &self, encoder_id: CommandEncoderId, label: &str ) -> Result<(), CommandEncoderError>

pub fn command_encoder_insert_debug_marker<A: HalApi>( &self, encoder_id: CommandEncoderId, label: &str ) -> Result<(), CommandEncoderError>

pub fn command_encoder_pop_debug_group<A: HalApi>( &self, encoder_id: CommandEncoderId ) -> Result<(), CommandEncoderError>

impl Global

pub fn adapter_is_surface_supported<A: HalApi>( &self, adapter_id: AdapterId, surface_id: SurfaceId ) -> Result<bool, IsSurfaceSupportedError>

pub fn surface_get_capabilities<A: HalApi>( &self, surface_id: SurfaceId, adapter_id: AdapterId ) -> Result<SurfaceCapabilities, GetSurfaceSupportError>

pub fn device_features<A: HalApi>( &self, device_id: DeviceId ) -> Result<Features, InvalidDevice>

pub fn device_limits<A: HalApi>( &self, device_id: DeviceId ) -> Result<Limits, InvalidDevice>

pub fn device_downlevel_properties<A: HalApi>( &self, device_id: DeviceId ) -> Result<DownlevelCapabilities, InvalidDevice>

pub fn device_create_buffer<A: HalApi>( &self, device_id: DeviceId, desc: &BufferDescriptor<'_>, id_in: Option<BufferId> ) -> (BufferId, Option<CreateBufferError>)

pub fn create_buffer_error<A: HalApi>( &self, id_in: Option<BufferId>, label: Label<'_> )

Assign id_in an error with the given label.

Ensure that future attempts to use id_in as a buffer ID will propagate the error, following the WebGPU “contagious invalidity” style.

Firefox uses this function to comply strictly with the WebGPU spec, which requires GPUBufferDescriptor validation to be generated on the Device timeline and leave the newly created GPUBuffer invalid.

Ideally, we would simply let device_create_buffer take care of all of this, but some errors must be detected before we can even construct a wgpu_types::BufferDescriptor to give it. For example, the WebGPU API allows a GPUBufferDescriptor’s usage property to be any WebIDL unsigned long value, but we can’t construct a wgpu_types::BufferUsages value from values with unassigned bits set. This means we must validate usage before we can call device_create_buffer.

When that validation fails, we must arrange for the buffer id to be considered invalid. This method provides the means to do so.

pub fn create_render_bundle_error<A: HalApi>( &self, id_in: Option<RenderBundleId>, label: Label<'_> )

pub fn create_texture_error<A: HalApi>( &self, id_in: Option<TextureId>, label: Label<'_> )

Assign id_in an error with the given label.

See create_buffer_error for more context and explanation.

pub fn device_wait_for_buffer<A: HalApi>( &self, device_id: DeviceId, buffer_id: BufferId ) -> Result<(), WaitIdleError>

Available on crate feature replay only.

pub fn buffer_label<A: HalApi>(&self, id: BufferId) -> String

pub fn buffer_destroy<A: HalApi>( &self, buffer_id: BufferId ) -> Result<(), DestroyError>

pub fn buffer_drop<A: HalApi>(&self, buffer_id: BufferId, wait: bool)

pub fn device_create_texture<A: HalApi>( &self, device_id: DeviceId, desc: &TextureDescriptor<'_>, id_in: Option<TextureId> ) -> (TextureId, Option<CreateTextureError>)

pub unsafe fn create_texture_from_hal<A: HalApi>( &self, hal_texture: A::Texture, device_id: DeviceId, desc: &TextureDescriptor<'_>, id_in: Option<TextureId> ) -> (TextureId, Option<CreateTextureError>)

Safety

hal_texture must be created from device_id corresponding raw handle.

hal_texture must be created respecting desc

hal_texture must be initialized

pub unsafe fn create_buffer_from_hal<A: HalApi>( &self, hal_buffer: A::Buffer, device_id: DeviceId, desc: &BufferDescriptor<'_>, id_in: Option<BufferId> ) -> (BufferId, Option<CreateBufferError>)

Safety

hal_buffer must be created from device_id corresponding raw handle.

hal_buffer must be created respecting desc

hal_buffer must be initialized

pub fn texture_label<A: HalApi>(&self, id: TextureId) -> String

pub fn texture_destroy<A: HalApi>( &self, texture_id: TextureId ) -> Result<(), DestroyError>

pub fn texture_drop<A: HalApi>(&self, texture_id: TextureId, wait: bool)

pub fn texture_create_view<A: HalApi>( &self, texture_id: TextureId, desc: &TextureViewDescriptor<'_>, id_in: Option<TextureViewId> ) -> (TextureViewId, Option<CreateTextureViewError>)

pub fn texture_view_label<A: HalApi>(&self, id: TextureViewId) -> String

pub fn texture_view_drop<A: HalApi>( &self, texture_view_id: TextureViewId, wait: bool ) -> Result<(), TextureViewDestroyError>

pub fn device_create_sampler<A: HalApi>( &self, device_id: DeviceId, desc: &SamplerDescriptor<'_>, id_in: Option<SamplerId> ) -> (SamplerId, Option<CreateSamplerError>)

pub fn sampler_label<A: HalApi>(&self, id: SamplerId) -> String

pub fn sampler_drop<A: HalApi>(&self, sampler_id: SamplerId)

pub fn device_create_bind_group_layout<A: HalApi>( &self, device_id: DeviceId, desc: &BindGroupLayoutDescriptor<'_>, id_in: Option<BindGroupLayoutId> ) -> (BindGroupLayoutId, Option<CreateBindGroupLayoutError>)

pub fn bind_group_layout_label<A: HalApi>( &self, id: BindGroupLayoutId ) -> String

pub fn bind_group_layout_drop<A: HalApi>( &self, bind_group_layout_id: BindGroupLayoutId )

pub fn device_create_pipeline_layout<A: HalApi>( &self, device_id: DeviceId, desc: &PipelineLayoutDescriptor<'_>, id_in: Option<PipelineLayoutId> ) -> (PipelineLayoutId, Option<CreatePipelineLayoutError>)

pub fn pipeline_layout_label<A: HalApi>(&self, id: PipelineLayoutId) -> String

pub fn pipeline_layout_drop<A: HalApi>( &self, pipeline_layout_id: PipelineLayoutId )

pub fn device_create_bind_group<A: HalApi>( &self, device_id: DeviceId, desc: &BindGroupDescriptor<'_>, id_in: Option<BindGroupId> ) -> (BindGroupId, Option<CreateBindGroupError>)

pub fn bind_group_label<A: HalApi>(&self, id: BindGroupId) -> String

pub fn bind_group_drop<A: HalApi>(&self, bind_group_id: BindGroupId)

pub fn device_create_shader_module<A: HalApi>( &self, device_id: DeviceId, desc: &ShaderModuleDescriptor<'_>, source: ShaderModuleSource<'_>, id_in: Option<ShaderModuleId> ) -> (ShaderModuleId, Option<CreateShaderModuleError>)

Create a shader module with the given source.

This function may consume a lot of stack space. Compiler-enforced limits for parsing recursion exist; if shader compilation runs into them, it will return an error gracefully. However, on some build profiles and platforms, the default stack size for a thread may be exceeded before this limit is reached during parsing. Callers should ensure that there is enough stack space for this, particularly if calls to this method are exposed to user input.

pub unsafe fn device_create_shader_module_spirv<A: HalApi>( &self, device_id: DeviceId, desc: &ShaderModuleDescriptor<'_>, source: Cow<'_, [u32]>, id_in: Option<ShaderModuleId> ) -> (ShaderModuleId, Option<CreateShaderModuleError>)

Safety

This function passes SPIR-V binary to the backend as-is and can potentially result in a driver crash.

pub fn shader_module_label<A: HalApi>(&self, id: ShaderModuleId) -> String

pub fn shader_module_drop<A: HalApi>(&self, shader_module_id: ShaderModuleId)

pub fn device_create_command_encoder<A: HalApi>( &self, device_id: DeviceId, desc: &CommandEncoderDescriptor<Label<'_>>, id_in: Option<CommandEncoderId> ) -> (CommandEncoderId, Option<DeviceError>)

pub fn command_buffer_label<A: HalApi>(&self, id: CommandBufferId) -> String

pub fn command_encoder_drop<A: HalApi>( &self, command_encoder_id: CommandEncoderId )

pub fn command_buffer_drop<A: HalApi>(&self, command_buffer_id: CommandBufferId)

pub fn device_create_render_bundle_encoder( &self, device_id: DeviceId, desc: &RenderBundleEncoderDescriptor<'_> ) -> (*mut RenderBundleEncoder, Option<CreateRenderBundleError>)

pub fn render_bundle_encoder_finish<A: HalApi>( &self, bundle_encoder: RenderBundleEncoder, desc: &RenderBundleDescriptor<'_>, id_in: Option<RenderBundleId> ) -> (RenderBundleId, Option<RenderBundleError>)

pub fn render_bundle_label<A: HalApi>(&self, id: RenderBundleId) -> String

pub fn render_bundle_drop<A: HalApi>(&self, render_bundle_id: RenderBundleId)

pub fn device_create_query_set<A: HalApi>( &self, device_id: DeviceId, desc: &QuerySetDescriptor<'_>, id_in: Option<QuerySetId> ) -> (QuerySetId, Option<CreateQuerySetError>)

pub fn query_set_drop<A: HalApi>(&self, query_set_id: QuerySetId)

pub fn query_set_label<A: HalApi>(&self, id: QuerySetId) -> String

pub fn device_create_render_pipeline<A: HalApi>( &self, device_id: DeviceId, desc: &RenderPipelineDescriptor<'_>, id_in: Option<RenderPipelineId>, implicit_pipeline_ids: Option<ImplicitPipelineIds<'_>> ) -> (RenderPipelineId, Option<CreateRenderPipelineError>)

pub fn render_pipeline_get_bind_group_layout<A: HalApi>( &self, pipeline_id: RenderPipelineId, index: u32, id_in: Option<BindGroupLayoutId> ) -> (BindGroupLayoutId, Option<GetBindGroupLayoutError>)

Get an ID of one of the bind group layouts. The ID adds a refcount, which needs to be released by calling bind_group_layout_drop.

pub fn render_pipeline_label<A: HalApi>(&self, id: RenderPipelineId) -> String

pub fn render_pipeline_drop<A: HalApi>( &self, render_pipeline_id: RenderPipelineId )

pub fn device_create_compute_pipeline<A: HalApi>( &self, device_id: DeviceId, desc: &ComputePipelineDescriptor<'_>, id_in: Option<ComputePipelineId>, implicit_pipeline_ids: Option<ImplicitPipelineIds<'_>> ) -> (ComputePipelineId, Option<CreateComputePipelineError>)

pub fn compute_pipeline_get_bind_group_layout<A: HalApi>( &self, pipeline_id: ComputePipelineId, index: u32, id_in: Option<BindGroupLayoutId> ) -> (BindGroupLayoutId, Option<GetBindGroupLayoutError>)

Get an ID of one of the bind group layouts. The ID adds a refcount, which needs to be released by calling bind_group_layout_drop.

pub fn compute_pipeline_label<A: HalApi>(&self, id: ComputePipelineId) -> String

pub fn compute_pipeline_drop<A: HalApi>( &self, compute_pipeline_id: ComputePipelineId )

pub fn surface_configure<A: HalApi>( &self, surface_id: SurfaceId, device_id: DeviceId, config: &SurfaceConfiguration<Vec<TextureFormat>> ) -> Option<ConfigureSurfaceError>

pub fn device_maintain_ids<A: HalApi>( &self, device_id: DeviceId ) -> Result<(), InvalidDevice>

Available on crate feature replay only.

Only triangle suspected resource IDs. This helps us to avoid ID collisions upon creating new resources when re-playing a trace.

pub fn device_poll<A: HalApi>( &self, device_id: DeviceId, maintain: Maintain<WrappedSubmissionIndex> ) -> Result<bool, WaitIdleError>

Check device_id for freeable resources and completed buffer mappings.

Return queue_empty indicating whether there are more queue submissions still in flight.

pub fn poll_all_devices(&self, force_wait: bool) -> Result<bool, WaitIdleError>

Poll all devices on all backends.

This is the implementation of wgpu::Instance::poll_all.

Return all_queue_empty indicating whether there are more queue submissions still in flight.

pub fn device_label<A: HalApi>(&self, id: DeviceId) -> String

pub fn device_start_capture<A: HalApi>(&self, id: DeviceId)

pub fn device_stop_capture<A: HalApi>(&self, id: DeviceId)

pub fn device_make_invalid<A: HalApi>(&self, device_id: DeviceId)

pub fn device_drop<A: HalApi>(&self, device_id: DeviceId)

pub fn device_set_device_lost_closure<A: HalApi>( &self, device_id: DeviceId, device_lost_closure: DeviceLostClosure )

pub fn device_destroy<A: HalApi>(&self, device_id: DeviceId)

pub fn device_mark_lost<A: HalApi>(&self, device_id: DeviceId, message: &str)

pub fn queue_drop<A: HalApi>(&self, queue_id: QueueId)

pub fn buffer_map_async<A: HalApi>( &self, buffer_id: BufferId, offset: BufferAddress, size: Option<BufferAddress>, op: BufferMapOperation ) -> BufferAccessResult

pub fn buffer_get_mapped_range<A: HalApi>( &self, buffer_id: BufferId, offset: BufferAddress, size: Option<BufferAddress> ) -> Result<(*mut u8, u64), BufferAccessError>

pub fn buffer_unmap<A: HalApi>(&self, buffer_id: BufferId) -> BufferAccessResult

impl Global

pub fn queue_write_buffer<A: HalApi>( &self, queue_id: QueueId, buffer_id: BufferId, buffer_offset: BufferAddress, data: &[u8] ) -> Result<(), QueueWriteError>

pub fn queue_create_staging_buffer<A: HalApi>( &self, queue_id: QueueId, buffer_size: BufferSize, id_in: Option<StagingBufferId> ) -> Result<(StagingBufferId, *mut u8), QueueWriteError>

pub fn queue_write_staging_buffer<A: HalApi>( &self, queue_id: QueueId, buffer_id: BufferId, buffer_offset: BufferAddress, staging_buffer_id: StagingBufferId ) -> Result<(), QueueWriteError>

pub fn queue_validate_write_buffer<A: HalApi>( &self, _queue_id: QueueId, buffer_id: BufferId, buffer_offset: u64, buffer_size: u64 ) -> Result<(), QueueWriteError>

pub fn queue_write_texture<A: HalApi>( &self, queue_id: QueueId, destination: &ImageCopyTexture, data: &[u8], data_layout: &ImageDataLayout, size: &Extent3d ) -> Result<(), QueueWriteError>

pub fn queue_submit<A: HalApi>( &self, queue_id: QueueId, command_buffer_ids: &[CommandBufferId] ) -> Result<WrappedSubmissionIndex, QueueSubmitError>

pub fn queue_get_timestamp_period<A: HalApi>( &self, queue_id: QueueId ) -> Result<f32, InvalidQueue>

pub fn queue_on_submitted_work_done<A: HalApi>( &self, queue_id: QueueId, closure: SubmittedWorkDoneClosure ) -> Result<(), InvalidQueue>

impl Global

pub fn new(name: &str, instance_desc: InstanceDescriptor) -> Self

pub unsafe fn from_hal_instance<A: HalApi>( name: &str, hal_instance: A::Instance ) -> Self

Safety

Refer to the creation of wgpu-hal Instance for every backend.

pub unsafe fn instance_as_hal<A: HalApi>(&self) -> Option<&A::Instance>

Safety

The raw instance handle returned must not be manually destroyed.

pub unsafe fn from_instance(instance: Instance) -> Self

Safety

The raw handles obtained from the Instance must not be manually destroyed

pub fn clear_backend<A: HalApi>(&self, _dummy: ())

pub fn generate_report(&self) -> GlobalReport

impl Global

pub unsafe fn instance_create_surface( &self, display_handle: RawDisplayHandle, window_handle: RawWindowHandle, id_in: Option<SurfaceId> ) -> Result<SurfaceId, CreateSurfaceError>

Available on crate feature raw-window-handle only.

Safety

display_handle must be a valid object to create a surface upon.

window_handle must remain valid as long as the returned SurfaceId is being used.

pub fn surface_drop(&self, id: SurfaceId)

pub fn enumerate_adapters( &self, inputs: AdapterInputs<'_, Adapter> ) -> Vec<AdapterId>

pub fn request_adapter( &self, desc: &RequestAdapterOptions, inputs: AdapterInputs<'_, Adapter> ) -> Result<AdapterId, RequestAdapterError>

pub unsafe fn create_adapter_from_hal<A: HalApi>( &self, hal_adapter: ExposedAdapter<A>, input: Option<AdapterId> ) -> AdapterId

Safety

hal_adapter must be created from this global internal instance handle.

pub fn adapter_get_info<A: HalApi>( &self, adapter_id: AdapterId ) -> Result<AdapterInfo, InvalidAdapter>

pub fn adapter_get_texture_format_features<A: HalApi>( &self, adapter_id: AdapterId, format: TextureFormat ) -> Result<TextureFormatFeatures, InvalidAdapter>

pub fn adapter_features<A: HalApi>( &self, adapter_id: AdapterId ) -> Result<Features, InvalidAdapter>

pub fn adapter_limits<A: HalApi>( &self, adapter_id: AdapterId ) -> Result<Limits, InvalidAdapter>

pub fn adapter_downlevel_capabilities<A: HalApi>( &self, adapter_id: AdapterId ) -> Result<DownlevelCapabilities, InvalidAdapter>

pub fn adapter_get_presentation_timestamp<A: HalApi>( &self, adapter_id: AdapterId ) -> Result<PresentationTimestamp, InvalidAdapter>

pub fn adapter_drop<A: HalApi>(&self, adapter_id: AdapterId)

impl Global

pub fn adapter_request_device<A: HalApi>( &self, adapter_id: AdapterId, desc: &DeviceDescriptor<'_>, trace_path: Option<&Path>, device_id_in: Option<DeviceId>, queue_id_in: Option<QueueId> ) -> (DeviceId, QueueId, Option<RequestDeviceError>)

pub unsafe fn create_device_from_hal<A: HalApi>( &self, adapter_id: AdapterId, hal_device: OpenDevice<A>, desc: &DeviceDescriptor<'_>, trace_path: Option<&Path>, device_id_in: Option<DeviceId>, queue_id_in: Option<QueueId> ) -> (DeviceId, QueueId, Option<RequestDeviceError>)

Safety

hal_device must be created from adapter_id or its internal handle.

desc must be a subset of hal_device features and limits.

impl Global

pub fn surface_get_current_texture<A: HalApi>( &self, surface_id: SurfaceId, texture_id_in: Option<TextureId> ) -> Result<SurfaceOutput, SurfaceError>

pub fn surface_present<A: HalApi>( &self, surface_id: SurfaceId ) -> Result<Status, SurfaceError>

pub fn surface_texture_discard<A: HalApi>( &self, surface_id: SurfaceId ) -> Result<(), SurfaceError>

impl Global

pub unsafe fn texture_as_hal<A: HalApi, F: FnOnce(Option<&A::Texture>) -> R, R>( &self, id: TextureId, hal_texture_callback: F ) -> R

Safety

The raw texture handle must not be manually destroyed

pub unsafe fn texture_view_as_hal<A: HalApi, F: FnOnce(Option<&A::TextureView>) -> R, R>( &self, id: TextureViewId, hal_texture_view_callback: F ) -> R

Safety

The raw texture view handle must not be manually destroyed

pub unsafe fn adapter_as_hal<A: HalApi, F: FnOnce(Option<&A::Adapter>) -> R, R>( &self, id: AdapterId, hal_adapter_callback: F ) -> R

Safety

The raw adapter handle must not be manually destroyed

pub unsafe fn device_as_hal<A: HalApi, F: FnOnce(Option<&A::Device>) -> R, R>( &self, id: DeviceId, hal_device_callback: F ) -> R

Safety

The raw device handle must not be manually destroyed

pub unsafe fn device_fence_as_hal<A: HalApi, F: FnOnce(Option<&A::Fence>) -> R, R>( &self, id: DeviceId, hal_fence_callback: F ) -> R

Safety

The raw fence handle must not be manually destroyed

pub unsafe fn surface_as_hal<A: HalApi, F: FnOnce(Option<&A::Surface>) -> R, R>( &self, id: SurfaceId, hal_surface_callback: F ) -> R

Safety

The raw surface handle must not be manually destroyed

pub unsafe fn command_encoder_as_hal_mut<A: HalApi, F: FnOnce(Option<&mut A::CommandEncoder>) -> R, R>( &self, id: CommandEncoderId, hal_command_encoder_callback: F ) -> R

Safety

The raw command encoder handle must not be manually destroyed