Module wgpu_core::hub

Structs

Global

Hub

Hubs

IdentityManager
A simple structure to manage identities of objects.

IdentityManagerFactory

Registry

Storage

Token
A permission token to lock resource T or anything after it, as defined by the Access implementations.

Enums

Root

Traits

Access
Type system for enforcing the lock order on shared HUB structures. If type A implements Access<B>, that means we are allowed to proceed with locking resource B after we lock A.

GfxBackend

GlobalIdentityHandlerFactory

IdentityHandler

IdentityHandlerFactory

Type Definitions

Input