Struct wgpu_core::hub::Global

pub struct Global<G: GlobalIdentityHandlerFactory> {
    pub instance: Instance,
    pub surfaces: Registry<Surface, SurfaceId, G>,
    hubs: Hubs<G>,
}

Fields
instance: Instancesurfaces: Registry<Surface, SurfaceId, G>hubs: Hubs<G>
Implementations

impl<G: GlobalIdentityHandlerFactory> Global<G>

pub fn render_bundle_encoder_finish<B: GfxBackend>(
&self,
bundle_encoder: RenderBundleEncoder,
desc: &RenderBundleDescriptor<'_>,
id_in: Input<G, RenderBundleId>
) -> Result<RenderBundleId, RenderBundleError>

impl<G: GlobalIdentityHandlerFactory> Global<G>

pub fn command_encoder_run_compute_pass<B: GfxBackend>(
&self,
encoder_id: CommandEncoderId,
pass: &ComputePass
) -> Result<(), ComputePassError>

impl<G: GlobalIdentityHandlerFactory> Global<G>

pub fn command_encoder_run_render_pass<B: GfxBackend>(
&self,
encoder_id: CommandEncoderId,
pass: &RenderPass
) -> Result<(), RenderPassError>

impl<G: GlobalIdentityHandlerFactory> Global<G>

pub fn command_encoder_copy_buffer_to_buffer<B: GfxBackend>(
&self,
command_encoder_id: CommandEncoderId,
source: BufferId,
source_offset: BufferAddress,
destination: BufferId,
destination_offset: BufferAddress,
size: BufferAddress
) -> Result<(), CopyError>

pub fn command_encoder_copy_buffer_to_texture<B: GfxBackend>(
&self,
command_encoder_id: CommandEncoderId,
source: &BufferCopyView,
destination: &TextureCopyView,
copy_size: &Extent3d
) -> Result<(), CopyError>

pub fn command_encoder_copy_texture_to_buffer<B: GfxBackend>(
&self,
command_encoder_id: CommandEncoderId,
source: &TextureCopyView,
destination: &BufferCopyView,
copy_size: &Extent3d
) -> Result<(), CopyError>

pub fn command_encoder_copy_texture_to_texture<B: GfxBackend>(
&self,
command_encoder_id: CommandEncoderId,
source: &TextureCopyView,
destination: &TextureCopyView,
copy_size: &Extent3d
) -> Result<(), CopyError>

impl<G: GlobalIdentityHandlerFactory> Global<G>

pub fn command_encoder_finish<B: GfxBackend>(
&self,
encoder_id: CommandEncoderId,
_desc: &CommandBufferDescriptor<Label<'_>>
) -> Result<CommandBufferId, CommandEncoderError>

pub fn command_encoder_push_debug_group<B: GfxBackend>(
&self,
encoder_id: CommandEncoderId,
label: &str
) -> Result<(), CommandEncoderError>

pub fn command_encoder_insert_debug_marker<B: GfxBackend>(
&self,
encoder_id: CommandEncoderId,
label: &str
) -> Result<(), CommandEncoderError>

pub fn command_encoder_pop_debug_group<B: GfxBackend>(
&self,
encoder_id: CommandEncoderId
) -> Result<(), CommandEncoderError>

impl<G: GlobalIdentityHandlerFactory> Global<G>

pub fn queue_write_buffer<B: GfxBackend>(
&self,
queue_id: QueueId,
buffer_id: BufferId,
buffer_offset: BufferAddress,
data: &[u8]
) -> Result<(), QueueWriteError>

pub fn queue_write_texture<B: GfxBackend>(
&self,
queue_id: QueueId,
destination: &TextureCopyView,
data: &[u8],
data_layout: &TextureDataLayout,
size: &Extent3d
) -> Result<(), QueueWriteError>

pub fn queue_submit<B: GfxBackend>(
&self,
queue_id: QueueId,
command_buffer_ids: &[CommandBufferId]
) -> Result<(), QueueSubmitError>

impl<G: GlobalIdentityHandlerFactory> Global<G>

pub fn device_features<B: GfxBackend>(
&self,
device_id: DeviceId
) -> Result<Features, InvalidDevice>

pub fn device_limits<B: GfxBackend>(
&self,
device_id: DeviceId
) -> Result<Limits, InvalidDevice>

pub fn device_create_buffer<B: GfxBackend>(
&self,
device_id: DeviceId,
desc: &BufferDescriptor<'_>,
id_in: Input<G, BufferId>
) -> Result<BufferId, CreateBufferError>

pub fn device_wait_for_buffer<B: GfxBackend>(
&self,
device_id: DeviceId,
buffer_id: BufferId
) -> Result<(), WaitIdleError>

pub fn device_set_buffer_sub_data<B: GfxBackend>(
&self,
device_id: DeviceId,
buffer_id: BufferId,
offset: BufferAddress,
data: &[u8]
) -> Result<(), BufferAccessError>

pub fn device_get_buffer_sub_data<B: GfxBackend>(
&self,
device_id: DeviceId,
buffer_id: BufferId,
offset: BufferAddress,
data: &mut [u8]
) -> Result<(), BufferAccessError>

pub fn buffer_error<B: GfxBackend>(&self, id_in: Input<G, BufferId>) -> BufferId

pub fn buffer_drop<B: GfxBackend>(&self, buffer_id: BufferId, now: bool)

pub fn device_create_texture<B: GfxBackend>(
&self,
device_id: DeviceId,
desc: &TextureDescriptor<'_>,
id_in: Input<G, TextureId>
) -> Result<TextureId, CreateTextureError>

pub fn texture_error<B: GfxBackend>(
&self,
id_in: Input<G, TextureId>
) -> TextureId

pub fn texture_drop<B: GfxBackend>(&self, texture_id: TextureId)

pub fn texture_create_view<B: GfxBackend>(
&self,
texture_id: TextureId,
desc: &TextureViewDescriptor<'_>,
id_in: Input<G, TextureViewId>
) -> Result<TextureViewId, CreateTextureViewError>

pub fn texture_view_error<B: GfxBackend>(
&self,
id_in: Input<G, TextureViewId>
) -> TextureViewId

pub fn texture_view_drop<B: GfxBackend>(
&self,
texture_view_id: TextureViewId
) -> Result<(), TextureViewDestroyError>

pub fn device_create_sampler<B: GfxBackend>(
&self,
device_id: DeviceId,
desc: &SamplerDescriptor<'_>,
id_in: Input<G, SamplerId>
) -> Result<SamplerId, CreateSamplerError>

pub fn sampler_error<B: GfxBackend>(
&self,
id_in: Input<G, SamplerId>
) -> SamplerId

pub fn sampler_drop<B: GfxBackend>(&self, sampler_id: SamplerId)

pub fn device_create_bind_group_layout<B: GfxBackend>(
&self,
device_id: DeviceId,
desc: &BindGroupLayoutDescriptor<'_>,
id_in: Input<G, BindGroupLayoutId>
) -> Result<BindGroupLayoutId, CreateBindGroupLayoutError>

pub fn bind_group_layout_error<B: GfxBackend>(
&self,
id_in: Input<G, BindGroupLayoutId>
) -> BindGroupLayoutId

pub fn bind_group_layout_drop<B: GfxBackend>(
&self,
bind_group_layout_id: BindGroupLayoutId
)

pub fn device_create_pipeline_layout<B: GfxBackend>(
&self,
device_id: DeviceId,
desc: &PipelineLayoutDescriptor<'_>,
id_in: Input<G, PipelineLayoutId>
) -> Result<PipelineLayoutId, CreatePipelineLayoutError>

pub fn pipeline_layout_error<B: GfxBackend>(
&self,
id_in: Input<G, PipelineLayoutId>
) -> PipelineLayoutId

pub fn pipeline_layout_drop<B: GfxBackend>(
&self,
pipeline_layout_id: PipelineLayoutId
)

fn derive_pipeline_layout<B: GfxBackend>(
&self,
device: &Device<B>,
device_id: DeviceId,
implicit_pipeline_ids: Option<ImplicitPipelineIds<'_, G>>,
derived_group_layouts: ArrayVec<[HashMap<u32, BindGroupLayoutEntry, BuildHasherDefault<FxHasher>>; 8]>,
bgl_guard: &mut Storage<BindGroupLayout<B>, BindGroupLayoutId>,
pipeline_layout_guard: &mut Storage<PipelineLayout<B>, PipelineLayoutId>
) -> Result<(PipelineLayoutId, ImplicitBindGroupCount), ImplicitLayoutError>

pub fn device_create_bind_group<B: GfxBackend>(
&self,
device_id: DeviceId,
desc: &BindGroupDescriptor<'_>,
id_in: Input<G, BindGroupId>
) -> Result<BindGroupId, CreateBindGroupError>

pub fn bind_group_error<B: GfxBackend>(
&self,
id_in: Input<G, BindGroupId>
) -> BindGroupId

pub fn bind_group_drop<B: GfxBackend>(&self, bind_group_id: BindGroupId)

pub fn device_create_shader_module<B: GfxBackend>(
&self,
device_id: DeviceId,
source: ShaderModuleSource<'_>,
id_in: Input<G, ShaderModuleId>
) -> Result<ShaderModuleId, CreateShaderModuleError>

pub fn shader_module_error<B: GfxBackend>(
&self,
id_in: Input<G, ShaderModuleId>
) -> ShaderModuleId

pub fn shader_module_drop<B: GfxBackend>(
&self,
shader_module_id: ShaderModuleId
)

pub fn device_create_command_encoder<B: GfxBackend>(
&self,
device_id: DeviceId,
desc: &CommandEncoderDescriptor<Label<'_>>,
id_in: Input<G, CommandEncoderId>
) -> Result<CommandEncoderId, CommandAllocatorError>

pub fn command_encoder_error<B: GfxBackend>(
&self,
id_in: Input<G, CommandEncoderId>
) -> CommandEncoderId

pub fn command_buffer_error<B: GfxBackend>(
&self,
id_in: Input<G, CommandBufferId>
) -> CommandBufferId

pub fn command_encoder_drop<B: GfxBackend>(
&self,
command_encoder_id: CommandEncoderId
)

pub fn command_buffer_drop<B: GfxBackend>(
&self,
command_buffer_id: CommandBufferId
)

pub fn device_create_render_bundle_encoder(
&self,
device_id: DeviceId,
desc: &RenderBundleEncoderDescriptor<'_>
) -> Result<RenderBundleEncoderId, CreateRenderBundleError>

pub fn render_bundle_error<B: GfxBackend>(
&self,
id_in: Input<G, RenderBundleId>
) -> RenderBundleId

pub fn render_bundle_drop<B: GfxBackend>(
&self,
render_bundle_id: RenderBundleId
)

pub fn device_create_render_pipeline<B: GfxBackend>(
&self,
device_id: DeviceId,
desc: &RenderPipelineDescriptor<'_>,
id_in: Input<G, RenderPipelineId>,
implicit_pipeline_ids: Option<ImplicitPipelineIds<'_, G>>
) -> Result<(RenderPipelineId, ImplicitBindGroupCount), CreateRenderPipelineError>

pub fn render_pipeline_get_bind_group_layout<B: GfxBackend>(
&self,
pipeline_id: RenderPipelineId,
index: u32
) -> Result<BindGroupLayoutId, GetBindGroupLayoutError>

Get an ID of one of the bind group layouts. The ID adds a refcount, which needs to be released by calling bind_group_layout_drop.

pub fn render_pipeline_error<B: GfxBackend>(
&self,
id_in: Input<G, RenderPipelineId>
) -> RenderPipelineId

pub fn render_pipeline_drop<B: GfxBackend>(
&self,
render_pipeline_id: RenderPipelineId
)

pub fn device_create_compute_pipeline<B: GfxBackend>(
&self,
device_id: DeviceId,
desc: &ComputePipelineDescriptor<'_>,
id_in: Input<G, ComputePipelineId>,
implicit_pipeline_ids: Option<ImplicitPipelineIds<'_, G>>
) -> Result<(ComputePipelineId, ImplicitBindGroupCount), CreateComputePipelineError>

pub fn compute_pipeline_get_bind_group_layout<B: GfxBackend>(
&self,
pipeline_id: ComputePipelineId,
index: u32
) -> Result<BindGroupLayoutId, GetBindGroupLayoutError>

Get an ID of one of the bind group layouts. The ID adds a refcount, which needs to be released by calling bind_group_layout_drop.

pub fn compute_pipeline_error<B: GfxBackend>(
&self,
id_in: Input<G, ComputePipelineId>
) -> ComputePipelineId

pub fn compute_pipeline_drop<B: GfxBackend>(
&self,
compute_pipeline_id: ComputePipelineId
)

pub fn device_get_swap_chain_preferred_format<B: GfxBackend>(
&self,
_device_id: DeviceId
) -> Result<TextureFormat, InvalidDevice>

pub fn device_create_swap_chain<B: GfxBackend>(
&self,
device_id: DeviceId,
surface_id: SurfaceId,
desc: &SwapChainDescriptor
) -> Result<SwapChainId, CreateSwapChainError>

pub fn device_maintain_ids<B: GfxBackend>(
&self,
device_id: DeviceId
) -> Result<(), InvalidDevice>

Only triange suspected resource IDs. This helps us to avoid ID collisions upon creating new resources when re-playing a trace.

pub fn device_poll<B: GfxBackend>(
&self,
device_id: DeviceId,
force_wait: bool
) -> Result<(), WaitIdleError>

fn poll_devices<B: GfxBackend>(
&self,
force_wait: bool,
callbacks: &mut Vec<(BufferMapOperation, BufferMapAsyncStatus)>
) -> Result<(), WaitIdleError>

pub fn poll_all_devices(&self, force_wait: bool) -> Result<(), WaitIdleError>

pub fn device_error<B: GfxBackend>(&self, id_in: Input<G, DeviceId>) -> DeviceId

pub fn device_drop<B: GfxBackend>(&self, device_id: DeviceId)

pub fn buffer_map_async<B: GfxBackend>(
&self,
buffer_id: BufferId,
range: Range<BufferAddress>,
op: BufferMapOperation
) -> Result<(), BufferAccessError>

pub fn buffer_get_mapped_range<B: GfxBackend>(
&self,
buffer_id: BufferId,
offset: BufferAddress,
_size: Option<BufferSize>
) -> Result<*mut u8, BufferAccessError>

pub fn buffer_unmap<B: GfxBackend>(
&self,
buffer_id: BufferId
) -> Result<(), BufferAccessError>

impl<G: GlobalIdentityHandlerFactory> Global<G>

pub fn new(name: &str, factory: G, backends: BackendBit) -> Self

pub fn clear_backend<B: GfxBackend>(&self, _dummy: ())

impl<G: GlobalIdentityHandlerFactory> Global<G>

pub fn enumerate_adapters(
&self,
inputs: AdapterInputs<'_, Input<G, AdapterId>>
) -> Vec<AdapterId>

pub fn request_adapter(
&self,
desc: &RequestAdapterOptions,
inputs: AdapterInputs<'_, Input<G, AdapterId>>
) -> Result<AdapterId, RequestAdapterError>

pub fn adapter_get_info<B: GfxBackend>(
&self,
adapter_id: AdapterId
) -> Result<AdapterInfo, InvalidAdapter>

pub fn adapter_features<B: GfxBackend>(
&self,
adapter_id: AdapterId
) -> Result<Features, InvalidAdapter>

pub fn adapter_limits<B: GfxBackend>(
&self,
adapter_id: AdapterId
) -> Result<Limits, InvalidAdapter>

pub fn adapter_destroy<B: GfxBackend>(&self, adapter_id: AdapterId)

impl<G: GlobalIdentityHandlerFactory> Global<G>

pub fn adapter_request_device<B: GfxBackend>(
&self,
adapter_id: AdapterId,
desc: &DeviceDescriptor,
trace_path: Option<&Path>,
id_in: Input<G, DeviceId>
) -> Result<DeviceId, RequestDeviceError>

impl<G: GlobalIdentityHandlerFactory> Global<G>

pub fn swap_chain_get_current_texture_view<B: GfxBackend>(
&self,
swap_chain_id: SwapChainId,
view_id_in: Input<G, TextureViewId>
) -> Result<SwapChainOutput, SwapChainError>

pub fn swap_chain_present<B: GfxBackend>(
&self,
swap_chain_id: SwapChainId
) -> Result<SwapChainStatus, SwapChainError>