Trait wgpu_core::hub::Access

pub trait Access<A> { }

Type system for enforcing the lock order on shared HUB structures. If type A implements Access<A>, that means we are allowed to proceed with locking resource B after we lock A.

The implenentations basically describe the edges in a directed graph of lock transitions. As long as it doesn’t have loops, we can have multiple concurrent paths on this graph (from multiple threads) without deadlocks, i.e. there is always a path whose next resource is not locked by some other path, at any time.