wgpu_core::id
Trait Marker

pub trait Marker: 'static + WasmNotSendSync { }

Marker trait used to determine which types uniquely identify a resource.

For example, Device<A> will have the same type of identifier as Device<B> because Device<T> for any T defines the same maker type.