wgpu_core::instance
Struct Instance

pub struct Instance {
    name: String,
    instance_per_backend: Vec<(Backend, Box<dyn DynInstance>)>,
    requested_backends: Backends,
    supported_backends: Backends,
    pub flags: InstanceFlags,
}

Fields
name: Stringinstance_per_backend: Vec<(Backend, Box<dyn DynInstance>)>

List of instances per wgpu-hal backend.

The ordering in this list implies prioritization and needs to be preserved.

requested_backends: Backends

The backends that were requested by the user.

supported_backends: Backends

The backends that we could have attempted to obtain from wgpu-hal — those for which support is compiled in, currently.

The union of this and requested_backends is the set of backends that would be used, independent of whether accessing the drivers/hardware for them succeeds. To obtain the set of backends actually in use by this instance, check instance_per_backend instead.

flags: InstanceFlags
Implementations

impl Instance

pub fn new(name: &str, instance_desc: &InstanceDescriptor) -> Self

fn try_add_hal<A: HalApi>(&mut self, _: A, instance_desc: &InstanceDescriptor)

Helper for Instance::new(); attempts to add a single wgpu-hal backend to this instance.

pub(crate) fn from_hal_instance<A: HalApi>( name: String, hal_instance: <A as Api>::Instance, ) -> Self

pub fn raw(&self, backend: Backend) -> Option<&dyn DynInstance>

pub unsafe fn as_hal<A: HalApi>(&self) -> Option<&A::Instance>

Safety

The raw instance handle returned must not be manually destroyed.

pub unsafe fn create_surface( &self, display_handle: RawDisplayHandle, window_handle: RawWindowHandle, ) -> Result<Surface, CreateSurfaceError>

Available on crate feature raw-window-handle only.

Creates a new surface targeting the given display/window handles.

Internally attempts to create hal surfaces for all enabled backends.

Fails only if creation for surfaces for all enabled backends fails in which case the error for each enabled backend is listed. Vice versa, if creation for any backend succeeds, success is returned. Surface creation errors are logged to the debug log in any case.

Safety

display_handle must be a valid object to create a surface upon.

window_handle must remain valid as long as the returned SurfaceId is being used.

pub unsafe fn create_surface_from_drm( &self, fd: i32, plane: u32, connector_id: u32, width: u32, height: u32, refresh_rate: u32, ) -> Result<Surface, CreateSurfaceError>

Available on Unix and non-Apple and non-target_family="wasm" only.

Creates a new surface from the given drm configuration.

Safety

All parameters must point to valid DRM values.

Platform Support

This function is only available on non-apple Unix-like platforms (Linux, FreeBSD) and currently only works with the Vulkan backend.

pub fn enumerate_adapters(&self, backends: Backends) -> Vec<Adapter>

pub fn request_adapter( &self, desc: &RequestAdapterOptions<&Surface>, backends: Backends, ) -> Result<Adapter, RequestAdapterError>

fn active_backends(&self) -> Backends