Module wgpu_core::lock::rank

Ranks for wgpu-core locks, restricting acquisition order.

See LockRank.

Macros

define_lock_ranks 🔒

Define a set of lock ranks, and each rank’s permitted successors.

Structs

LockRank

The rank of a lock.

LockRankSet

A bitflags type representing a set of lock ranks.

Enums

LockRankNumber 🔒

Constants

BUFFER_BIND_GROUPS

BUFFER_INITIALIZATION_STATUS

BUFFER_MAP_STATE

COMMAND_ALLOCATOR_FREE_ENCODERS

COMMAND_BUFFER_DATA

DEVICE_DEFERRED_DESTROY

DEVICE_FENCE

DEVICE_LIFE_TRACKER

DEVICE_PENDING_WRITES

DEVICE_SNATCHABLE_LOCK

DEVICE_TRACE

DEVICE_TRACKERS

DEVICE_USAGE_SCOPES

IDENTITY_MANAGER_VALUES

REGISTRY_STORAGE

RESOURCE_POOL_INNER

SHARED_TRACKER_INDEX_ALLOCATOR_INNER

SURFACE_PRESENTATION

TEXTURE_BIND_GROUPS

TEXTURE_INITIALIZATION_STATUS

TEXTURE_VIEWS