wgpu_core
Module resource

Structs

Blas

Buffer

BufferMapOperation

BufferPendingMapping 🔒

DestroyedBuffer

A buffer that has been marked as destroyed and is staged for actual deletion soon.

DestroyedResourceError

DestroyedTexture

A texture that has been marked as destroyed and is staged for actual deletion soon.

FlushedStagingBuffer

HalTextureViewDescriptor 🔒

InvalidResourceError

MissingBufferUsageError

MissingTextureUsageError

QuerySet

ResourceErrorIdent

Sampler

SamplerDescriptor

Describes a Sampler

StagingBuffer

A temporary buffer, consumed by the command that uses it.

Texture

TextureView

TextureViewDescriptor

Describes a TextureView.

Tlas

TrackingData 🔒

Information about the wgpu-core resource.

Enums

BufferAccessError

BufferMapState 🔒

CreateBufferError

CreateQuerySetError

CreateSamplerError

CreateTextureError

CreateTextureViewError

DestroyError

Fallible

SamplerFilterErrorType

TextureClearMode

TextureDimensionError

TextureErrorDimension

TextureInner 🔒

TextureViewDestroyError

TextureViewNotRenderableReason

Traits

AccelerationStructure 🔒

Labeled

ParentDevice

ResourceType

Trackable 🔒

Type Aliases

BlasDescriptor

BufferAccessResult

BufferDescriptor

BufferMapCallbacksend_sync

QuerySetDescriptor

TextureDescriptor

TlasDescriptor