wgpu_core

Module resource

Structs

Blas

BlasPendingCompact 🔒

Buffer

BufferMapOperation

BufferPendingMapping 🔒

DestroyedBuffer

A buffer that has been marked as destroyed and is staged for actual deletion soon.

DestroyedResourceError

DestroyedTexture

A texture that has been marked as destroyed and is staged for actual deletion soon.

ExternalTexture

FlushedStagingBuffer

HalTextureViewDescriptor 🔒

InvalidResourceError

MissingBufferUsageError

MissingTextureUsageError

QuerySet

ResourceErrorIdent

Sampler

SamplerDescriptor

Describes a Sampler

StagingBuffer

A temporary buffer, consumed by the command that uses it.

Texture

TextureView

TextureViewDescriptor

Describes a TextureView.

Tlas

TrackingData 🔒

Information about the wgpu-core resource.

Enums

BlasCompactState 🔒

BufferAccessError

BufferMapState 🔒

CreateBufferError

CreateExternalTextureError

CreateQuerySetError

CreateSamplerError

CreateTextureError

CreateTextureViewError

Fallible

SamplerFilterErrorType

TextureClearMode

TextureDimensionError

TextureErrorDimension

TextureInner 🔒

TextureViewDestroyError

TextureViewNotRenderableReason

Traits

Labeled

ParentDevice

RawResourceAccess

Allow access to the hal resource as guarded by the SnatchGuard.

ResourceType

Trackable 🔒

Type Aliases

BlasCompactCallbacksend_sync

BlasDescriptor

BlasPrepareCompactResult

BufferAccessResult

BufferDescriptor

BufferMapCallbacksend_sync

ExternalTextureDescriptor

QuerySetDescriptor

TextureDescriptor

TlasDescriptor