Module wgpu_core::resource

Structs

Buffer

BufferMapCallback

BufferMapCallbackC

BufferMapOperation

QuerySet

ResourceInfo

Information about the wgpu-core resource.

Sampler

SamplerDescriptor

Describes a Sampler

StagingBuffer

A temporary buffer, consumed by the command that uses it.

Texture

TextureView

TextureViewDescriptor

Describes a TextureView.

Enums

BufferAccessError

BufferMapAsyncStatus

The status code provided to the buffer mapping callback.

CreateBufferError

CreateQuerySetError

CreateSamplerError

CreateTextureError

CreateTextureViewError

DestroyError

SamplerFilterErrorType

TextureClearMode

TextureDimensionError

TextureErrorDimension

TextureViewDestroyError

TextureViewNotRenderableReason

Traits

Resource

Type Aliases

BufferAccessResult

BufferDescriptor

QuerySetDescriptor

TextureDescriptor