wgpu_core::timestamp_normalization

Struct TimestampNormalizer

pub struct TimestampNormalizer {
    state: Option<InternalState>,
}

Normalizes GPU timestamps to have a consistent 1GHz period. See module documentation for more information.

Fields
state: Option<InternalState>
Implementations

impl TimestampNormalizer

pub fn new( device: &Device, timestamp_period: f32, ) -> Result<Self, TimestampNormalizerInitError>

Creates a new timestamp normalizer.

If the device cannot support automatic timestamp normalization, this will return a normalizer that does nothing.

Errors

If any resources are invalid, this will return an error.

pub unsafe fn create_normalization_bind_group( &self, device: &Device, buffer: &dyn DynBuffer, buffer_label: Option<&str>, buffer_size: BufferSize, buffer_usages: BufferUsages, ) -> Result<TimestampNormalizationBindGroup, DeviceError>

Create a bind group for normalizing timestamps in buffer.

This function is unsafe because it does not know that buffer_size is the true size of the buffer.

pub fn normalize( &self, snatch_guard: &SnatchGuard<'_>, encoder: &mut dyn DynCommandEncoder, tracker: &mut BufferTracker, bind_group: &TimestampNormalizationBindGroup, buffer: &Arc<Buffer>, buffer_offset_bytes: u64, total_timestamps: u32, )

pub fn dispose(self, device: &dyn DynDevice)

pub fn enabled(&self) -> bool