Module wgpu_core::track

Structs

PendingTransition
A structure containing all the information about a particular resource transition. User code should be able to generate a pipeline barrier based on the contents.

ResourceTracker
A tracker for all resources of a given type.

TextureState

TrackerSet
A set of trackers for all relevant resources.

Unit
A single unit of state tracking. It keeps an initial usage as well as the last/current one, similar to Range.

Constants

DUMMY_SELECTOR

Traits

ResourceState
The main trait that abstracts away the tracking logic of a particular resource type, like a buffer or a texture.

Type Definitions

BufferState