wgpu_core::track
Module buffer

Buffer Trackers

Buffers are represented by a single state for the whole resource, a 16 bit bitflag of buffer usages. Because there is only ever one subresource, they have no selector.

Structs

BufferBindGroupState 🔒

Stores a bind group’s buffers + their usages (within the bind group).

BufferTracker 🔒

Stores all buffer state within a command buffer.

BufferUsageScope 🔒

Stores all buffer state within a single usage scope.

DeviceBufferTracker 🔒

Stores all buffer state within a device.

Enums

BufferStateProvider 🔒

Source of Buffer State.

Functions

barrier 🔒 ⚠

insert 🔒 ⚠

merge 🔒 ⚠

update 🔒 ⚠