wgpu_core::track

Module metadata

The ResourceMetadata type.

Structs

ResourceMetadata 🔒

A set of resources, holding a Arc<T> and epoch for each member.

Enums

ResourceMetadataProvider 🔒

A source of resource metadata.

Functions

iterate_bitvec_indices 🔒

Produces an iterator that yields the indexes of all bits that are set in the bitvec.

resize_bitvec 🔒

Resizes the given bitvec to the given size. I’m not sure why this is hard to do but it is.