Struct wgpu_core::track::ResourceTracker

pub(crate) struct ResourceTracker<S: ResourceState> {
    map: HashMap<u32, Resource<S>, BuildHasherDefault<FxHasher>>,
    temp: Vec<PendingTransition<S>>,
    backend: Backend,
}

A tracker for all resources of a given type.

Fields
map: HashMap<u32, Resource<S>, BuildHasherDefault<FxHasher>>

An association of known resource indices with their tracked states.

temp: Vec<PendingTransition<S>>

Temporary storage for collecting transitions.

backend: Backend

The backend variant for all the tracked resources.

Implementations

impl<S: ResourceState> ResourceTracker<S>

pub fn new(backend: Backend) -> Self

Create a new empty tracker.

pub(crate) fn remove(&mut self, id: Valid<S::Id>) -> bool

Remove an id from the tracked map.

pub(crate) fn remove_abandoned(&mut self, id: Valid<S::Id>) -> bool

Removes the resource from the tracker if we are holding the last reference.

pub(crate) fn optimize(&mut self)

Try to optimize the internal representation.

pub fn used<'a>(&'a self) -> impl 'a + Iterator<Item = Valid<S::Id>>

Return an iterator over used resources keys.

fn clear(&mut self)

Clear the tracked contents.

pub(crate) fn init(
&mut self,
id: Valid<S::Id>,
ref_count: RefCount,
state: S
) -> Result<(), &S>

Initialize a resource to be used.

Returns false if the resource is already registered.

pub fn query(&self, id: Valid<S::Id>, selector: S::Selector) -> Option<S::Usage>

Query the usage of a resource selector.

Returns Some(Usage) only if this usage is consistent across the given selector.

fn get_or_insert<'a>(
self_backend: Backend,
map: &'a mut HashMap<u32, Resource<S>, BuildHasherDefault<FxHasher>>,
id: Valid<S::Id>,
ref_count: &RefCount
) -> &'a mut Resource<S>

Make sure that a resource is tracked, and return a mutable reference to it.

pub(crate) fn change_extend(
&mut self,
id: Valid<S::Id>,
ref_count: &RefCount,
selector: S::Selector,
usage: S::Usage
) -> Result<(), PendingTransition<S>>

Extend the usage of a specified resource.

Returns conflicting transition as an error.

pub(crate) fn change_replace(
&mut self,
id: Valid<S::Id>,
ref_count: &RefCount,
selector: S::Selector,
usage: S::Usage
) -> Drain<'_, PendingTransition<S>>

Replace the usage of a specified resource.

pub(crate) fn prepend(
&mut self,
id: Valid<S::Id>,
ref_count: &RefCount,
selector: S::Selector,
usage: S::Usage
) -> Result<(), PendingTransition<S>>

Turn the tracking from the “expand” mode into the “replace” one, installing the selected usage as the “first”. This is a special operation only used by the render pass attachments.

pub(crate) fn merge_extend(
&mut self,
other: &Self
) -> Result<(), PendingTransition<S>>

Merge another tracker into self by extending the current states without any transitions.

pub(crate) fn merge_replace<'a>(
&'a mut self,
other: &'a Self
) -> Drain<'_, PendingTransition<S>>

Merge another tracker, adding it’s transitions to self. Transitions the current usage to the new one.

pub(crate) fn use_extend<'a, T: 'a + Borrow<RefCount>>(
&mut self,
storage: &'a Storage<T, S::Id>,
id: S::Id,
selector: S::Selector,
usage: S::Usage
) -> Result<&'a T, UseExtendError<S::Usage>>

Use a given resource provided by an Id with the specified usage. Combines storage access by ‘Id’ with the transition that extends the last read-only usage, if possible.

Returns the old usage as an error if there is a conflict.

pub(crate) fn use_replace<'a, T: 'a + Borrow<RefCount>>(
&mut self,
storage: &'a Storage<T, S::Id>,
id: S::Id,
selector: S::Selector,
usage: S::Usage
) -> Result<(&'a T, Drain<'_, PendingTransition<S>>), S::Id>

Use a given resource provided by an Id with the specified usage. Combines storage access by ‘Id’ with the transition that replaces the last usage with a new one, returning an iterator over these transitions.