wgpu_core::track
Trait ResourceUses

pub(crate) trait ResourceUses:
    Debug
    + BitAnd<Output = Self>
    + BitOr<Output = Self>
    + PartialEq
    + Sized
    + Copy {
    type Selector: Debug;

    const EXCLUSIVE: Self;

    // Required methods
    fn bits(self) -> u16;
    fn all_ordered(self) -> bool;
    fn any_exclusive(self) -> bool;
}

The uses that a resource or subresource can be in.

Required Associated Constants

const EXCLUSIVE: Self

All flags that are exclusive.

Required Associated Types

type Selector: Debug

The selector used by this resource.

Required Methods

fn bits(self) -> u16

Turn the resource into a pile of bits.

fn all_ordered(self) -> bool

Returns true if the all the uses are ordered.

fn any_exclusive(self) -> bool

Returns true if any of the uses are exclusive.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.