Trait wgpu_core::track::ResourceUses

pub(crate) trait ResourceUses: Debug + BitAnd<Output = Self> + BitOr<Output = Self> + PartialEq + Sized + Copy {
    type Id: Copy + Debug + TypedId;
    type Selector: Debug;

    const EXCLUSIVE: Self;

    // Required methods
    fn bits(self) -> u16;
    fn all_ordered(self) -> bool;
    fn any_exclusive(self) -> bool;
}

The uses that a resource or subresource can be in.

Required Associated Types

type Id: Copy + Debug + TypedId

The relevant resource ID type.

type Selector: Debug

The selector used by this resource.

Required Associated Constants

const EXCLUSIVE: Self

All flags that are exclusive.

Required Methods

fn bits(self) -> u16

Turn the resource into a pile of bits.

fn all_ordered(self) -> bool

Returns true if the all the uses are ordered.

fn any_exclusive(self) -> bool

Returns true if any of the uses are exclusive.