wgpu_core

Module validation

Structs

EntryPoint 🔒

EntryPointMeshInfo 🔒

Interface

InterfaceVar

NumericType

Resource 🔒

SpecializationConstant 🔒

StageIo

Enums

BindingError

BindingLayoutSource

BindingTypeName

FilteringError

InputError

NumericDimension 🔒

ResourceType 🔒

StageError

Errors produced when validating a programmable stage of a pipeline.

Varying 🔒

Functions

check_texture_format

Return true if the fragment format is covered by the provided output.

map_storage_format_from_naga

map_storage_format_to_naga

validate_color_attachment_bytes_per_sample

Validate a list of color attachment formats against maxColorAttachmentBytesPerSample.