Module wgpu_core::validation

Structs

MissingBufferUsageError

MissingTextureUsageError

Enums

BindingError

InputError

IntrospectionBindGroupLayouts

MaybeOwned

StageError

Errors produced when validating a programmable stage of a pipeline.

Functions

check_binding_use

check_buffer_usage

Checks that the given buffer usage contains the required buffer usage, returns an error otherwise.

check_stage

check_texture_format

Return true if the fragment format is covered by the provided output.

check_texture_usage

Checks that the given texture usage contains the required texture usage, returns an error otherwise.

derive_binding_type

get_aligned_type_size

is_sub_type

map_storage_format_from_naga

map_storage_format_to_naga

map_texture_format

map_vertex_format

Type Definitions

StageInterface